namada_sdk
Module address

Implements transparent addresses as described in Accounts Addresses.

Structs

EstablishedAddress

An established address is generated on-chain

EstablishedAddressGen

A generator of established addresses

ImplicitAddress

An implicit address is derived from a cryptographic key

Enums

Address

An account’s address

InternalAddress

An internal address represents a module with a native VP

Constants

ADDRESS_LEN

The length of Address encoded with Bech32m.

ESTABLISHED_ADDRESS_BYTES_LEN

The length of an established Address encoded with Borsh.

ETH_BRIDGE

Internal Eth bridge address

GOV

Internal Governance address

HASH_HEX_LEN

Length of a hash of an address as a hexadecimal string

HASH_LEN

Length of a trimmed hash of an address.

IBC

Internal IBC address

MASP

Internal MASP address

MULTITOKEN

Internal Multitoken address

PARAMETERS

Internal ledger parameters address

PGF

Internal Public Goods funding address

POS

Internal PoS address

POS_SLASH_POOL

Internal PoS slash pool address

SHA_HASH_LEN

SHA-256 hash len

TEMP_STORAGE

Address with temporary storage is used to pass data from txs to VPs which is never committed to DB

Functions

gen_deterministic_established_address

Generate a new established address. Unlike gen_established_address, this will give the same address for the same seed.

gen_established_address

Generate a new established address.

Type Aliases

DecodeError

Error from decoding address from string

Result

Result of decoding address from string