namada_sdk
Module args

Structures encapsulating SDK arguments

Structs

Bond

Bond arguments

BpConversionTableEntry

An entry in the Bridge pool recommendations conversion rates table.

BridgePoolProof

Bridge pool proof arguments.

BridgeValidatorSet

Bridge validator set arguments.

ClaimRewards

Claim arguments

CommissionRateChange

Commission rate change args

Complete

Generate shell completions

ConsensusKeyChange

Consensus key change args

Duration

Duration wrapper that provides a method to parse a value from a string.

EthereumBridgePool

A transfer to be added to the Ethereum bridge pool.

GenIbcShieldingTransfer

IBC shielding transfer generation arguments

GovernanceValidatorSet

Governance validator set arguments.

InitProposal

Transaction to initialize create a new proposal

KeyAddressAdd

Wallet key / address add arguments

KeyAddressFind

Wallet key / address lookup arguments

KeyAddressList

Wallet list arguments

KeyAddressRemove

Wallet key / address remove arguments

KeyConvert

Wallet key export arguments

KeyDerive

Wallet restore key and implicit address arguments

KeyExport

Wallet key export arguments

KeyGen

Wallet generate key and implicit address arguments

KeyImport

Wallet key import arguments

MetaDataChange

Commission rate change args

PayAddressGen

Generate payment address arguments

Query

Common query arguments

QueryAccount

Query token balance(s)

QueryBalance

Query token balance(s)

QueryBondedStake

Query PoS bonded stake

QueryBonds

Query PoS bond(s)

QueryCommissionRate

Query PoS commission rate

QueryConversions

Query asset conversions

QueryDelegations

Query PoS delegations

QueryEffNativeSupply

Query effective native supply

QueryFindValidator

Query PoS to find a validator

QueryMetaData

Query validator metadata

QueryPgf

Query pgf data

QueryProposal

Query proposal

QueryProposalVotes

Query proposal votes

QueryProtocolParameters

Query protocol parameters

QueryRawBytes

Query the raw bytes of given storage key

QueryResult

Transaction associated results arguments

QueryRewards

Query PoS rewards

QueryRewardsEstimate

Get an estimate for the MASP rewards accumulated by the next MASP epoch.

QuerySlashes

Query PoS slashes

QueryStakingRewardsRate

Query estimate of staking rewards rate

QueryTotalSupply

Query token total supply

QueryTransfers

Query historical transfer(s)

QueryValidatorState

Query the state of a validator (its validator set or if it is jailed)

QueryWithoutCtx

Common query arguments

RecommendBatch

Bridge pool batch recommendation.

Redelegate

Redelegation arguments

RelayBridgePoolProof

Arguments to an Ethereum Bridge pool relay operation.

ResignSteward

Commission rate change args

RevealPk

Reveal public key

SdkTypes

The concrete types being used in Namada SDK

ShieldedSync

Sync notes from MASP owned by the provided spending / viewing keys. Syncing can be told to stop at a given block height.

SignTx

Sign a transaction offline

Tx

Common transaction arguments

TxBecomeValidator

Transaction to initialize a new account

TxCustom

Custom transaction arguments

TxDeactivateValidator

Deactivate validator args

TxIbcTransfer

IBC transfer transaction arguments

TxInitAccount

Transaction to initialize a new account

TxInitValidator

Transaction to initialize a new account

TxReactivateValidator

Re-activate a deactivated validator args

TxShieldedTransfer

Shielded transfer transaction arguments

TxShieldedTransferData

Shielded transfer-specific arguments

TxShieldingTransfer

Shielding transfer transaction arguments

TxShieldingTransferData

Shielding transfer-specific arguments

TxTransparentTransfer

Transparent transfer transaction arguments

TxTransparentTransferData

Transparent transfer-specific arguments

TxUnjailValidator

Re-activate a jailed validator args

TxUnshieldingTransfer

Unshielding transfer transaction arguments

TxUnshieldingTransferData

Unshielding transfer-specific arguments

TxUpdateAccount

Transaction to update a VP arguments

Unbond

Unbond arguments

UpdateStewardCommission

Commission rate change args

ValidatorSetProof

Validator set proof arguments.

ValidatorSetUpdateRelay

Validator set update relayer arguments.

VoteProposal

Transaction to vote on a proposal

Withdraw

Withdraw arguments

Enums

DeviceTransport

Hardware Wallet transport - HID (USB) or TCP

InputAmount

An amount read in by the cli

Shell

Supported shell types

TxExpiration

The possible values for the tx expiration

Traits

NamadaTypes

Abstraction of types being used in Namada

TxBuilder

Builder functions for Tx