namada_sdk::args
Trait NamadaTypes

pub trait NamadaTypes: Clone + Debug {

    type Address: Clone + Debug;
    type AddrOrNativeToken: Clone + Debug + From<Self::Address>;
    type Keypair: Clone + Debug;
    type TendermintAddress: Clone + Debug;
    type ConfigRpcTendermintAddress: Clone + Debug + From<Self::TendermintAddress>;
    type EthereumAddress: Clone + Debug;
    type ViewingKey: Clone + Debug;
    type SpendingKey: Clone + Debug;
    type DatedViewingKey: Clone + Debug;
    type DatedSpendingKey: Clone + Debug;
    type PaymentAddress: Clone + Debug;
    type BalanceOwner: Clone + Debug;
    type PublicKey: Clone + Debug;
    type TransferSource: Clone + Debug;
    type TransferTarget: Clone + Debug;
    type Data: Clone + Debug;
    type BpConversionTable: Clone + Debug;
    type MaspIndexerAddress: Clone + Debug;
    type BlockHeight: Clone + Debug;

}

Abstraction of types being used in Namada

Required Associated Types

type Address: Clone + Debug

Represents an address on the ledger

type AddrOrNativeToken: Clone + Debug + From<Self::Address>

Represents an address that defaults to a native token

type Keypair: Clone + Debug

Represents a key pair

type TendermintAddress: Clone + Debug

Represents the address of a Tendermint endpoint (used in context-less CLI commands where chain config isn’t available)

type ConfigRpcTendermintAddress: Clone + Debug + From<Self::TendermintAddress>

RPC address of a locally configured node

type EthereumAddress: Clone + Debug

Represents the address of an Ethereum endpoint

type ViewingKey: Clone + Debug

Represents a shielded viewing key

type SpendingKey: Clone + Debug

Represents a shielded spending key

type DatedViewingKey: Clone + Debug

Represents a shielded viewing key

type DatedSpendingKey: Clone + Debug

Represents a shielded spending key

type PaymentAddress: Clone + Debug

Represents a shielded payment address

type BalanceOwner: Clone + Debug

Represents the owner of a balance

type PublicKey: Clone + Debug

Represents a public key

type TransferSource: Clone + Debug

Represents the source of a Transfer

type TransferTarget: Clone + Debug

Represents the target of a Transfer

type Data: Clone + Debug

Represents some data that is used in a transaction

type BpConversionTable: Clone + Debug

Bridge pool recommendations conversion rates table.

type MaspIndexerAddress: Clone + Debug

Address of a namada-masp-indexer live instance

type BlockHeight: Clone + Debug

Represents a block height

Object Safety

This trait is not object safe.