namada_sdk::args
Trait TxBuilder

pub trait TxBuilder<C: NamadaTypes>: Sized {

    // Required method
    fn tx<F>(self, func: F) -> Self
       where F: FnOnce(Tx<C>) -> Tx<C>;

    // Provided methods
    fn dry_run(self, dry_run: bool) -> Self { ... }
    fn dry_run_wrapper(self, dry_run_wrapper: bool) -> Self { ... }
    fn dump_tx(self, dump_tx: bool) -> Self { ... }
    fn output_folder(self, output_folder: PathBuf) -> Self { ... }
    fn force(self, force: bool) -> Self { ... }
    fn broadcast_only(self, broadcast_only: bool) -> Self { ... }
    fn ledger_address(self, ledger_address: C::TendermintAddress) -> Self { ... }
    fn initialized_account_alias(
        self,
        initialized_account_alias: String,
    ) -> Self { ... }
    fn wallet_alias_force(self, wallet_alias_force: bool) -> Self { ... }
    fn fee_amount(self, fee_amount: InputAmount) -> Self { ... }
    fn wrapper_fee_payer(self, wrapper_fee_payer: C::PublicKey) -> Self { ... }
    fn fee_token(self, fee_token: C::Address) -> Self { ... }
    fn gas_limit(self, gas_limit: GasLimit) -> Self { ... }
    fn expiration(self, expiration: TxExpiration) -> Self { ... }
    fn chain_id(self, chain_id: ChainId) -> Self { ... }
    fn signing_keys(self, signing_keys: Vec<C::PublicKey>) -> Self { ... }
    fn signatures(self, signatures: Vec<C::Data>) -> Self { ... }
    fn tx_reveal_code_path(self, tx_reveal_code_path: PathBuf) -> Self { ... }
    fn password(self, password: Zeroizing<String>) -> Self { ... }
    fn memo(self, memo: Vec<u8>) -> Self { ... }

}

Builder functions for Tx

Required Methods

fn tx<F>(self, func: F) -> Self
where F: FnOnce(Tx<C>) -> Tx<C>,

Apply the given function to the Tx inside self

Provided Methods

fn dry_run(self, dry_run: bool) -> Self

Simulate applying the transaction

fn dry_run_wrapper(self, dry_run_wrapper: bool) -> Self

Simulate applying both the wrapper and inner transactions

fn dump_tx(self, dump_tx: bool) -> Self

Dump the transaction bytes to file

fn output_folder(self, output_folder: PathBuf) -> Self

The output directory path to where serialize the data

fn force(self, force: bool) -> Self

Submit the transaction even if it doesn’t pass client checks

fn broadcast_only(self, broadcast_only: bool) -> Self

Do not wait for the transaction to be added to the blockchain

fn ledger_address(self, ledger_address: C::TendermintAddress) -> Self

The address of the ledger node as host:port

fn initialized_account_alias(self, initialized_account_alias: String) -> Self

If any new account is initialized by the tx, use the given alias to save it in the wallet.

fn wallet_alias_force(self, wallet_alias_force: bool) -> Self

Whether to force overwrite the above alias, if it is provided, in the wallet.

fn fee_amount(self, fee_amount: InputAmount) -> Self

The amount being paid (for gas unit) to include the transaction

fn wrapper_fee_payer(self, wrapper_fee_payer: C::PublicKey) -> Self

The fee payer signing key

fn fee_token(self, fee_token: C::Address) -> Self

The token in which the fee is being paid

fn gas_limit(self, gas_limit: GasLimit) -> Self

The max amount of gas used to process tx

fn expiration(self, expiration: TxExpiration) -> Self

The optional expiration of the transaction

fn chain_id(self, chain_id: ChainId) -> Self

The chain id for which the transaction is intended

fn signing_keys(self, signing_keys: Vec<C::PublicKey>) -> Self

Sign the tx with the key for the given alias from your wallet

fn signatures(self, signatures: Vec<C::Data>) -> Self

List of signatures to attach to the transaction

fn tx_reveal_code_path(self, tx_reveal_code_path: PathBuf) -> Self

Path to the TX WASM code file to reveal PK

fn password(self, password: Zeroizing<String>) -> Self

Password to decrypt key

fn memo(self, memo: Vec<u8>) -> Self

Change memo

Object Safety

This trait is not object safe.