namada_sdk::booleans
Trait ResultBoolExt

pub trait ResultBoolExt<E> {
    // Required methods
    fn true_or(self, error: E) -> Result<(), E>;
    fn true_or_else<F>(self, handle_err: F) -> Result<(), E>
       where F: FnOnce() -> E;
}

Extend Result of bool values with the possibility to create Result values of unit and some error type.

Required Methods

fn true_or(self, error: E) -> Result<(), E>

Return Ok(()) if Ok(true), Err(error) if Ok(false) or pass back the error if Err(_).

fn true_or_else<F>(self, handle_err: F) -> Result<(), E>
where F: FnOnce() -> E,

Return Ok(()) if Ok(true), Err(handle_err()) if Ok(false) or pass back the error if Err(_).

Object Safety

This trait is not object safe.