namada_sdk
Module borsh

Borsh binary encoding (re-exported) from official crate with custom ext.

Modules

de

error

io

Subset of std::io which is used as part of borsh public API.

schema

Module is available if borsh is built with features = ["unstable__schema"].

ser

Traits

BorshDeserialize

A data-structure that can be de-serialized from binary format by NBOR.

BorshSchema

The declaration and the definition of the type that can be used to (de)serialize Borsh without the Rust type that produced it.

BorshSerialize

A data-structure that can be serialized into binary format by NBOR.

BorshSerializeExt

Extensions to types implementing BorshSerialize.

Functions

from_reader

Deserializes an object from a reader.

from_slice

Deserializes an object from a slice of bytes.

max_serialized_size

Returns the largest possible size of a serialised object based solely on its type T.

object_length

Serializes an object without allocation to compute and return its length

schema_container_of

generate BorshSchemaContainer for type T

to_vec

Serialize an object into a vector of bytes.

to_writer

Serializes an object directly into a Writer.

try_from_slice_with_schema

Deserialize this instance from a slice of bytes, but assume that at the beginning we have bytes describing the schema of the type. We deserialize this schema and verify that it is correct.

try_to_vec_with_schema

Serialize object into a vector of bytes and prefix with the schema serialized as vector of bytes in Borsh format.

Derive Macros

BorshDeserialize

BorshSchema

BorshSerialize