namada_sdk::collections
Module hash_map

Modules

raw_entry_v1

Opt-in access to the experimental raw entry API.

serde_seq

Functions to serialize and deserialize an IndexMap as an ordered sequence.

Structs

Drain

A draining iterator over the entries of an IndexMap.

HashMap

A hash table where the iteration order of the key-value pairs is independent of the hash values of the keys.

IndexMap

A hash table where the iteration order of the key-value pairs is independent of the hash values of the keys.

IndexedEntry

A view into an occupied entry in an IndexMap obtained by index.

IntoIter

An owning iterator over the entries of an IndexMap.

IntoKeys

An owning iterator over the keys of an IndexMap.

IntoValues

An owning iterator over the values of an IndexMap.

Iter

An iterator over the entries of an IndexMap.

IterMut

A mutable iterator over the entries of an IndexMap.

IterMut2

A mutable iterator over the entries of an IndexMap.

Keys

An iterator over the keys of an IndexMap.

OccupiedEntry

A view into an occupied entry in an IndexMap. It is part of the Entry enum.

Slice

A dynamically-sized slice of key-value pairs in an IndexMap.

Splice

A splicing iterator for IndexMap.

VacantEntry

A view into a vacant entry in an IndexMap. It is part of the Entry enum.

Values

An iterator over the values of an IndexMap.

ValuesMut

A mutable iterator over the values of an IndexMap.

Enums

Entry

Entry for an existing key-value pair in an IndexMap or a vacant location to insert one.

Traits

MutableEntryKey

Opt-in mutable access to Entry keys.

MutableKeys

Opt-in mutable access to IndexMap keys.

RawEntryApiV1

Opt-in access to the experimental raw entry API.