namada_sdk::control_flow
Module time

Time related logic for futures.

Structs

Constant

Constant sleep strategy.

Duration

A Duration type to represent a span of time, typically used for system timeouts.

ExponentialBackoff

Exponential backoff sleep strategy.

Instant

A measurement of a monotonically nondecreasing clock. Opaque and useful only with Duration.

LinearBackoff

Linear backoff sleep strategy.

Map

Map a function to the duration returned from a sleep strategy.

RunForever

Run a fallible task forever.

RunWithRetries

A SleepRunUntil implementation, for running a fallible task a certain number of times before ultimately giving up.

Sleep

A SleepStrategy adaptor, to run async tasks with custom sleep durations.

Enums

Error

Future task related errors.

Traits

SleepRunUntil

Zero cost abstraction to check if we should exit a running SleepStrategy.

SleepStrategy

A sleep strategy to be applied to fallible runs of arbitrary tasks.

Functions

sleep

Pause the active task for the given duration.