namada_sdk::control_flow::time
Trait SleepRunUntil

pub trait SleepRunUntil {
    type Output<T>;

    // Required methods
    fn success<T>(self, ret: T) -> Self::Output<T>;
    fn error<T>(self) -> Self::Output<T>;
    fn poll_error(&mut self) -> bool;
}

Zero cost abstraction to check if we should exit a running SleepStrategy.

Required Associated Types

type Output<T>

The output type to return.

Required Methods

fn success<T>(self, ret: T) -> Self::Output<T>

Exit with success, returning a value.

Consumes the SleepRunUntil instance.

fn error<T>(self) -> Self::Output<T>

Exit with an error.

Consumes the SleepRunUntil instance.

fn poll_error(&mut self) -> bool

Check if an error has occurred, prompting an early exit.

Object Safety

This trait is not object safe.