namada_sdk::control_flow::time
Trait SleepStrategy

pub trait SleepStrategy {
    type State;

    // Required methods
    fn new_state() -> Self::State;
    fn backoff(&self, state: &Self::State) -> Duration;
    fn next_state(&self, state: &mut Self::State);
    fn prev_state(&self, state: &mut Self::State);

    // Provided method
    fn map<M>(self, map: M) -> Map<Self, M>
       where M: Fn(Duration) -> Duration,
             Self: Sized { ... }
}

A sleep strategy to be applied to fallible runs of arbitrary tasks.

Required Associated Types

type State

The state of the sleep strategy.

Required Methods

fn new_state() -> Self::State

Return a new sleep strategy state.

fn backoff(&self, state: &Self::State) -> Duration

Calculate a duration from a sleep strategy state.

fn next_state(&self, state: &mut Self::State)

Move to the next state of the sleep strategy.

fn prev_state(&self, state: &mut Self::State)

Move to the previous state of the sleep strategy.

Provided Methods

fn map<M>(self, map: M) -> Map<Self, M>
where M: Fn(Duration) -> Duration, Self: Sized,

Map a function to the duration returned from a sleep strategy.

Object Safety

This trait is not object safe.