namada_sdk
Module eth_bridge

Ethereum bridge utilities shared between wasm and the cli.

Re-exports

pub use ethers;

Modules

bridge_pool

Bridge pool SDK functionality.

event

Ethereum Bridge transaction events.

oracle

Ethereum bridge oracle

protocol

Ethereum bridge protocol transactions and validation.

storage

Functionality for accessing the storage subspace

structs

Ethereum bridge struct re-exports and types to do with ethereum.

token

Transparent token types, storage functions, and validation.

validator_set

Validator set updates SDK functionality.

vp

Native validity predicates for the Namada Ethereum bridge. This includes both the bridge vp and the vp for the bridge pool.

wrapped_erc20s

Functionality for accessing the multitoken subspace

Structs

BlockOnEthSync

Arguments to block_on_eth_sync.

ContractVersion

Represents a configuration value for the version of a contract that can be upgraded. Starts from 1.

Contracts

Represents all the Ethereum contracts that need to be directly know about by validators.

Erc20WhitelistEntry

An ERC20 token whitelist entry.

EthAssetMint

Number of tokens to mint after receiving a “transfer to Namada” Ethereum event.

EthBridgeQueriesHook

A handle to EthBridgeQueries.

EthereumBridgeParams

Represents chain parameters for the Ethereum bridge.

EthereumOracleConfig

Subset of EthereumBridgeParams, containing only Ethereum oracle specific parameters.

MinimumConfirmations

Represents a configuration value for the minimum number of confirmations an Ethereum event must reach before it can be acted on.

UpgradeableContract

Represents an Ethereum contract that may be upgraded.

Enums

EthBridgeEnabled

Enum indicating if the bridge was initialized at genesis or a later epoch.

EthBridgeStatus

An enum indicating if the Ethereum bridge is enabled.

SendValsetUpd

This enum is used as a parameter to EthBridgeQueriesHook::must_send_valset_upd.

SyncStatus

The result of querying an Ethereum nodes syncing status.

Constants

ADDRESS

Internal Eth bridge address

Traits

EthBridgeQueries

Methods used to query blockchain Ethereum bridge related state.

Functions

block_on_eth_sync

Block until Ethereum finishes synchronizing.

check_bridge_status

Check if the bridge is disabled, enabled, or scheduled to be enabled at a specified Epoch.

eth_sync_or

Check if Ethereum has finished synchronizing. In case it has not, perform action.

eth_sync_or_exit

Check if Ethereum has finished synchronizing. In case it has not, end execution.

eth_syncing_status

Fetch the sync status of an Ethereum node.

eth_syncing_status_timeout

Fetch the sync status of an Ethereum node, with a custom time out duration.

is_bridge_active_at

Returns a boolean indicating whether the bridge is currently active at the specified Epoch.

is_bridge_comptime_enabled

Check if the Ethereum Bridge has been enabled at compile time.

read_native_erc20_address

Get the Ethereum address for wNam from storage, if possible