namada_sdk::eth_bridge::protocol
Module transactions

This module contains functionality for handling protocol transactions.

When a protocol transaction is included in a block, we may expect all nodes to update their blockchain state in a deterministic way. This can be done natively rather than via the wasm environment as happens with regular transactions.

Modules

bridge_pool_roots

Functions dealing with bridge pool root hash.

ethereum_events

Code for handling Ethereum events protocol txs.

validator_set_update

Code for handling validator set update protocol txs.

votes

Logic and data types relating to tallying validators’ votes for pieces of data stored in the ledger, where those pieces of data should only be acted on once they have received enough votes

Type Aliases

ChangedKeys

The keys changed while applying a protocol transaction.