namada_sdk::eth_bridge::protocol::transactions::bridge_pool_roots
Function apply_derived_tx

pub fn apply_derived_tx<D, H, Gov>(
    state: &mut WlState<D, H>,
    vext: MultiSignedVext,
) -> Result<BatchedTxResult, Report>
where
    D: 'static + DB + for<'iter> DBIter<'iter> + Sync,
    H: 'static + StorageHasher + Sync,
    Gov: Read<WlState<D, H>>,

Applies a tally of signatures on over the Ethereum bridge pool root and nonce. Note that every signature passed into this function will be for the same root and nonce.

For roots + nonces which have been seen by a quorum of validators, the signature is made available for bridge pool proofs.