namada_sdk::eth_bridge::protocol::transactions::votes
Function dedupe

pub fn dedupe(
    signers: BTreeSet<(Address, BlockHeight)>,
) -> BTreeMap<Address, BlockHeight>

Deterministically constructs a Votes map from a set of validator addresses and the block heights they signed something at.

We arbitrarily take the earliest block height for each validator address encountered.