namada_sdk::eth_bridge::protocol::transactions::votes
Trait EpochedVotingPowerExt

pub trait EpochedVotingPowerExt {
    // Required methods
    fn epoch_max_voting_power<D, H, Gov>(
        &self,
        state: &WlState<D, H>,
    ) -> Option<Amount>
       where D: 'static + DB + for<'iter> DBIter<'iter> + Sync,
             H: 'static + StorageHasher + Sync,
             Gov: Read<WlState<D, H>>;
    fn tallied_stake(&self) -> Amount;

    // Provided methods
    fn fractional_stake<D, H, Gov>(
        &self,
        state: &WlState<D, H>,
    ) -> FractionalVotingPower
       where D: 'static + DB + for<'iter> DBIter<'iter> + Sync,
             H: 'static + StorageHasher + Sync,
             Gov: Read<WlState<D, H>> { ... }
    fn has_majority_quorum<D, H, Gov>(&self, state: &WlState<D, H>) -> bool
       where D: 'static + DB + for<'iter> DBIter<'iter> + Sync,
             H: 'static + StorageHasher + Sync,
             Gov: Read<WlState<D, H>> { ... }
}

Extension methods for EpochedVotingPower instances.

Required Methods

fn epoch_max_voting_power<D, H, Gov>( &self, state: &WlState<D, H>, ) -> Option<Amount>
where D: 'static + DB + for<'iter> DBIter<'iter> + Sync, H: 'static + StorageHasher + Sync, Gov: Read<WlState<D, H>>,

Query the stake of the most secure Epoch referenced by an EpochedVotingPower. This translates to the Epoch with the most staked tokens.

fn tallied_stake(&self) -> Amount

Fetch the sum of the stake tallied on an EpochedVotingPower.

Provided Methods

fn fractional_stake<D, H, Gov>( &self, state: &WlState<D, H>, ) -> FractionalVotingPower
where D: 'static + DB + for<'iter> DBIter<'iter> + Sync, H: 'static + StorageHasher + Sync, Gov: Read<WlState<D, H>>,

Fetch the sum of the stake tallied on an EpochedVotingPower, as a fraction over the maximum stake seen in the epochs voted on.

fn has_majority_quorum<D, H, Gov>(&self, state: &WlState<D, H>) -> bool
where D: 'static + DB + for<'iter> DBIter<'iter> + Sync, H: 'static + StorageHasher + Sync, Gov: Read<WlState<D, H>>,

Check if the Tally associated with an EpochedVotingPower can be considered seen.

Object Safety

This trait is not object safe.