namada_sdk::eth_bridge::protocol::transactions::votes
Type Alias Votes

pub type Votes = BTreeMap<Address, BlockHeight>;

The addresses of validators that voted for something, and the block heights at which they voted.

We use a BTreeMap to enforce that a validator (as uniquely identified by an Address) may vote at most once, their vote must be associated with a specific BlockHeight. Their voting power at that block height is what is used when calculating whether something has enough voting power behind it or not.

Aliased Type

struct Votes { /* private fields */ }