namada_sdk::eth_bridge::protocol::validation::validator_set_update
Function validate_valset_upd_vext

pub fn validate_valset_upd_vext<D, H, Gov>(
    state: &WlState<D, H>,
    ext: &SignedVext,
    last_epoch: Epoch,
) -> Result<(), VoteExtensionError>
where
    D: 'static + DB + for<'iter> DBIter<'iter>,
    H: 'static + StorageHasher,
    Gov: Read<WlState<D, H>>,

Validates a validator set update vote extension issued at the epoch provided as an argument.

Validation checks

To validate a validator_set_update::SignedVext, Namada nodes check if:

The signing validator is a consensus validator during the epoch signing_epoch inside the extension.

A validator set update proof is not available yet for signing_epoch.

The validator correctly signed the extension, with its Ethereum hot key.

The validator signed over the epoch inside of the extension, whose value should not be greater than last_epoch.

The voting powers in the vote extension correspond to the voting powers of the validators of signing_epoch + 1.

The voting powers signed over were Ethereum ABI encoded, normalized to 2^32, and sorted in descending order.