namada_sdk::eth_bridge
Module storage

Functionality for accessing the storage subspace

Modules

bridge_pool

Tools for accessing the storage subspaces of the Ethereum bridge pool

eth_bridge_queries

Storage queries for ethereum bridge.

parameters

Parameters for configuring the Ethereum bridge

proof

Proofs over some arbitrary data.

vote_tallies

Functionality for accessing keys to do with tallying votes

vp

Validity predicate storage

whitelist

ERC20 token whitelist storage data.

wrapped_erc20s

Functionality for accessing the multitoken subspace

Functions

active_key

A key for storing the active / inactive status of the Ethereum bridge.

bridge_contract_key

Storage key for the Ethereum address of the bridge contract.

escrow_key

The key to the escrow of the VP.

eth_start_height_key

Key for storing the initial Ethereum block height when events will first be extracted from.

has_eth_addr_segment

Check if the given key contains an Ethereum bridge address segment.

is_eth_bridge_key

Returns whether a key belongs to this account or not

min_confirmations_key

Storage key for the minimum confirmations parameter.

native_erc20_key

Storage key for the Ethereum address of wNam.

prefix

Key prefix for the storage subspace