namada_sdk::eth_bridge::storage
Module bridge_pool

Tools for accessing the storage subspaces of the Ethereum bridge pool

Structs

BridgePoolTree

A simple Merkle tree for the Ethereum bridge pool

Constants

BRIDGE_POOL_ADDRESS

The main address of the Ethereum bridge pool

Functions

get_key_from_hash

Get the storage key for the transfers using the hash

get_nonce_key

Get the storage key for the batch nonce of the bridge pool. Used for replay protection.

get_pending_key

Get the storage key for the transfers in the pool

get_signed_root_key

Get the storage key for the root of the Merkle tree containing the transfers in the pool

is_bridge_pool_key

Check if a key belongs to the bridge pools sub-storage

is_pending_transfer_key

Check if a key is for a pending transfer