namada_sdk::eth_bridge::storage
Module parameters

Parameters for configuring the Ethereum bridge

Structs

ContractVersion

Represents a configuration value for the version of a contract that can be upgraded. Starts from 1.

Contracts

Represents all the Ethereum contracts that need to be directly know about by validators.

Erc20WhitelistEntry

An ERC20 token whitelist entry.

EthereumBridgeParams

Represents chain parameters for the Ethereum bridge.

EthereumOracleConfig

Subset of EthereumBridgeParams, containing only Ethereum oracle specific parameters.

MinimumConfirmations

Represents a configuration value for the minimum number of confirmations an Ethereum event must reach before it can be acted on.

UpgradeableContract

Represents an Ethereum contract that may be upgraded.

Functions

read_native_erc20_address

Get the Ethereum address for wNam from storage, if possible