namada_sdk::eth_bridge
Module structs

Ethereum bridge struct re-exports and types to do with ethereum.

Structs

BlockHeight

This type must be able to represent any valid Ethereum block height. It must also be Borsh serializeable, so that it can be stored in blockchain storage.

ChainTransfer

ChainTransfer(uint256,address,string)

Erc20Transfer

Erc20Transfer(bytes32,uint256,address,address)

RelayProof

RelayProof((bytes32,uint256,address,address)[],bytes32,bytes32[],bool[],uint256,string)

Signature

Signature(bytes32,bytes32,uint8)

ValidatorSetArgs

ValidatorSetArgs(bytes32[],uint256)