namada_sdk::eth_bridge::token
Trait StorageWrite

pub trait StorageWrite {
    // Required methods
    fn write_bytes(
        &mut self,
        key: &Key,
        val: impl AsRef<[u8]>,
    ) -> Result<(), Error>;
    fn delete(&mut self, key: &Key) -> Result<(), Error>;

    // Provided methods
    fn write<T>(&mut self, key: &Key, val: T) -> Result<(), Error>
       where T: BorshSerialize { ... }
    fn delete_prefix(&mut self, prefix: &Key) -> Result<(), Error>
       where Self: Sized + StorageRead { ... }
}

Common storage write interface

Required Methods

fn write_bytes(&mut self, key: &Key, val: impl AsRef<[u8]>) -> Result<(), Error>

Write a value as bytes at the given key to storage.

fn delete(&mut self, key: &Key) -> Result<(), Error>

Delete a value at the given key from storage.

Provided Methods

fn write<T>(&mut self, key: &Key, val: T) -> Result<(), Error>
where T: BorshSerialize,

Write a value to be encoded with Borsh at the given key to storage.

fn delete_prefix(&mut self, prefix: &Key) -> Result<(), Error>
where Self: Sized + StorageRead,

Delete all key-vals with a matching prefix.

Object Safety

This trait is not object safe.