namada_sdk::eth_bridge::token::tx
Trait CreditOrDebit

pub trait CreditOrDebit {
    // Required methods
    fn keys(&self) -> impl Iterator<Item = (Address, Address)>;
    fn get(&self, key: &(Address, Address)) -> Option<&Amount>;
    fn into_iter(self) -> impl IntoIterator<Item = ((Address, Address), Amount)>;
}

Multi-transfer credit or debit amounts

Required Methods

fn keys(&self) -> impl Iterator<Item = (Address, Address)>

Gets an iterator over the pair of credited or debited owners and token addresses, in sorted order.

fn get(&self, key: &(Address, Address)) -> Option<&Amount>

Returns a reference to the value corresponding to pair of owner and token address

fn into_iter(self) -> impl IntoIterator<Item = ((Address, Address), Amount)>

Gets an owning iterator over the pairs of credited or debited owners and token addresses paired with the token amount, sorted by key.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.