namada_sdk::eth_bridge
Module validator_set

Validator set updates SDK functionality.

Functions

query_bridge_validator_set

Query an ABI encoding of the Bridge validator set at a given epoch.

query_governnace_validator_set

Query an ABI encoding of the Governance validator set at a given epoch.

query_validator_set_update_proof

Query an ABI encoding of the validator set to be installed at the given epoch, and its associated proof.

relay_validator_set_update

Relay a validator set update, signed off for a given epoch.