namada_sdk
Module events

Logic to do with events emitted by the ledger.

Modules

extend

Extend events with additional fields.

log

A log to store events emitted by FinalizeBlock calls in the ledger.

Macros

event_type

Instantiate a new EventType in const contexts. Mostly useful to define new event types in the protocol.

Structs

Attributes

A thin wrapper around a HashMap for parsing event JSONs returned in tendermint subscription responses.

Event

Custom events that can be queried from Tendermint using a websocket client

EventType

ABCI event type.

EventTypeBuilder

Build an EventType segment by segment.

Enums

EventError

Errors to do with emitting events.

EventLevel

Indicates if an event is emitted do to an individual Tx or the nature of a finalized block

Traits

EmitEvents

Used in sub-systems that may emit events.

EventToEmit

An event to be emitted in Namada.