namada_sdk::events::extend
Trait EventAttributeEntry

pub trait EventAttributeEntry<'a> {
    type Value;
    type ValueOwned;

    const KEY: &'static str;

    // Required method
    fn into_value(self) -> Self::Value;
}

Represents an entry in the attributes of an Event.

Required Associated Constants

const KEY: &'static str

Key to read or write and event attribute to.

Required Associated Types

type Value

Data to be stored in the given KEY.

type ValueOwned

Identical to Self::Value, with the exception that this should be an owned variant of that type.

Required Methods

fn into_value(self) -> Self::Value

Return the data to be stored in the given KEY.

Object Safety

This trait is not object safe.