namada_sdk::events
Module log

A log to store events emitted by FinalizeBlock calls in the ledger.

The log will hold up to N events of a certain kind at a time, before resorting to pruning older events contained within.

Modules

dumb_queries

Silly simple event matcher.

Structs

EventLog

Represents a log of Event instances emitted by FinalizeBlock calls, in the ledger.

Params

Parameters to configure the pruning of the event log.

WithMatcher

Iterator over an EventLog taking a matcher in order to filter events within.