namada_sdk
Module keccak

This module is for hashing Namada types using the keccak256 hash function in a way that is compatible with smart contracts on Ethereum.

Structs

Keccak

The Keccak hash functions defined in Keccak SHA3 submission.

KeccakHash

Represents a Keccak hash.

Enums

TryFromError

Errors for converting / parsing Keccak hashes

Traits

Hasher

A trait for hashing an arbitrary stream of bytes.

Functions

keccak_hash

Hash bytes using Keccak