namada_sdk
Module key

Cryptographic keys

Modules

common

Cryptographic keys

ed25519

Ed25519 keys and related functionality

secp256k1

secp256k1 keys and related functionality

Structs

PublicKeyHash

Public key hash derived from common::Key borsh encoded bytes (hex string of the first 40 chars of sha256 hash)

SerializeWithBorsh

Tag type that indicates we should use BorshSerialize to sign data in a Signable wrapper.

SignableEthMessage

Tag type that indicates we should use ABI serialization to sign data in a Signable wrapper.

Enums

ParsePublicKeyError

ParseSecretKeyError

ParseSignatureError

PkhFromStringError

SchemeType

Type capturing signature scheme IDs

VerifySigError

Represents an error in signature verification

Traits

PublicKey

Represents a public key

PublicKeyTmRawHash

Derive Tendermint raw hash from the public key

RefTo

A value-to-value conversion that consumes the input value.

SecretKey

Represents a secret key

SigScheme

Represents a digital signature scheme.

Signable

A serialization method to provide to namada_tx::Signed, such that we may sign serialized data.

SignableBytes

Helper trait to compress arbitrary bytes to a hash value, which can be signed over.

Signature

Represents a signature

TryFromRef

Simple and safe type conversions that may fail in a controlled way under some circumstances.

Functions

tm_consensus_key_raw_hash

Convert validator’s consensus key into address raw hash that is compatible with Tendermint

tm_raw_hash_to_string

Convert Tendermint validator’s raw hash bytes to Namada raw hash string