namada_sdk::key
Trait TryFromRef

pub trait TryFromRef<T>: Sized
where
    T: ?Sized,
{
    type Error;

    // Required method
    fn try_from_ref(value: &T) -> Result<Self, Self::Error>;
}

Simple and safe type conversions that may fail in a controlled way under some circumstances.

Required Associated Types

type Error

The type returned in the event of a conversion error.

Required Methods

fn try_from_ref(value: &T) -> Result<Self, Self::Error>

Performs the conversion.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.