namada_sdk
Macro router

macro_rules! router {
    { $name:ident, $( $pattern:tt $( -> $return_type:path )? = $handle:tt , )* } => { ... };
}

Compile time tree patterns router with type-safe dynamic parameter parsing, automatic routing, type-safe path constructors and optional client query methods.

The router! macro implements greedy matching algorithm.

Examples

router! {ROOT,
  // This pattern matches `/pattern_a/something`, where `something` can be
  // parsed with `FromStr` into `ArgType`.
  ( "pattern_a" / [typed_dynamic_arg: ArgType] ) -> ReturnType = handler,

  // Optional arguments.
  ( "pattern_b" / [optional_dynamic_arg: opt ArgType] ) -> ReturnType =
    handler,

  // NOTE: In paths optional arguments that have `None` value are skipped.
  // Because of this if there are two consecutive `opt` args of the same
  // type it is ambiguous which one is specified if only one has a value.
  // To work-around this, simply insert a literal pattern in between the
  // args (`"separator"`) like:
  ( "pattern_c" / [arg1: opt ArgType] / "separator" / [arg2: opt ArgType] ) -> ReturnType =
    handler,

  // Untyped dynamic arg is a string slice `&str`
  ( "pattern_d" / [untyped_dynamic_arg] ) -> ReturnType = handler,

  // The handler additionally receives the `RequestQuery`, which can have
  // some data attached, specified block height and ask for a proof. It
  // returns `EncodedResponseQuery` (the `data` must be encoded, if
  // necessary), which can have some `info` string and a proof.
  ( "pattern_e" ) -> ReturnType = (with_options handler),

  ( "another" / "pattern" / "that" / "goes" / "deep" ) -> ReturnType = handler,

  // Inlined sub-tree
  ( "subtree" / [this_is_fine: ArgType] ) = {
    ( "a" ) -> u64 = a_handler,
    ( "b" / [another_arg] ) -> u64 = b_handler,
  }

  // Imported sub-router - The prefix can only have literal segments
  ( "sub" / "no_dynamic_args" ) = (sub SUB_ROUTER),
}

router! {SUB_ROUTER,
  ( "pattern" ) -> ReturnType = handler,
}

Handler functions used in the patterns should have the expected signature:

fn handler<D, H>(ctx: RequestCtx<'_, D, H>, args ...)
  -> namada_storage::Result<ReturnType>
where
    D: 'static + DB + for<'iter> DBIter<'iter> + Sync,
    H: 'static + StorageHasher + Sync;

If the handler wants to support request options, it can be defined as (with_options $handler) and then the expected signature is:

fn handler<D, H>(ctx: RequestCtx<'_, D, H>, request: &RequestQuery, args
...)   -> namada_storage::Result<ResponseQuery<ReturnType>>
where
    D: 'static + DB + for<'iter> DBIter<'iter> + Sync,
    H: 'static + StorageHasher + Sync;