namada_sdk::masp
Module utils

Helper functions and types

Structs

Fetched

A cache of fetched indexed transactions.

TrialDecrypted

Cache of decrypted notes.

Enums

MaspClientCapabilities

Enumerates the capabilities of a MaspClient implementation.

RetryStrategy

When retrying to fetch all notes in a loop, this dictates the strategy for how many attempts should be made.

Traits

MaspClient

This abstracts away the implementation details of how shielded-sync fetches the necessary data from a remote server.

Functions

blocks_left_to_fetch

Given a block height range we wish to request and a cache of fetched block heights, returns the set of sub-ranges we need to request so that all blocks in the inclusive range [from, to] get cached.

Type Aliases

DecryptedData

Type alias for a successful note decryption.

IndexedNoteData

Type alias for convenience and profit

IndexedNoteEntry

Type alias for the entries of IndexedNoteData iterators

IndexedNoteEntryRefs

Borrowed version of an IndexedNoteEntry