namada_sdk
Crate masp_primitives

General MASP primitives.

masp_primitives is a library that provides the core structs and functions necessary for working with MASP based on Zcash Sapling.

Modules

asset_type

bls12_381

bls12_381

consensus

Consensus logic and parameters.

constants

Various constants used by the Zcash primitives.

convert

ff

This crate provides traits for working with finite fields.

group

jubjub

This crate provides an implementation of the Jubjub elliptic curve and its associated field arithmetic. See README.md for more details about Jubjub.

keys

memo

Structs for handling encrypted memos.

merkle_tree

Implementation of a Merkle tree of commitments used to prove the existence of notes.

num_traits

Numeric traits for generic mathematics

sapling

Structs and constants specific to the Sapling shielded pool.

transaction

zip32

Implementation of ZIP 32 for hierarchical deterministic key management.

Traits

MaybeArbitrary