namada_sdk::masp_primitives::num_traits::ops
Module checked

Traits

CheckedAdd

Performs addition, returning None if overflow occurred.

CheckedDiv

Performs division, returning None on division by zero or if overflow occurred.

CheckedMul

Performs multiplication, returning None if overflow occurred.

CheckedNeg

Performs negation, returning None if the result can’t be represented.

CheckedRem

Performs integral remainder, returning None on division by zero or if overflow occurred.

CheckedShl

Performs shift left, returning None on shifts larger than or equal to the type width.

CheckedShr

Performs shift right, returning None on shifts larger than or equal to the type width.

CheckedSub

Performs subtraction, returning None if overflow occurred.