namada_sdk::masp_primitives::num_traits::ops::checked
Trait CheckedShl

pub trait CheckedShl<Rhs = u32>: Sized {
    type Output;

    // Required method
    fn checked_shl(self, rhs: Rhs) -> Option<Self::Output>;
}

Performs shift left, returning None on shifts larger than or equal to the type width.

Required Associated Types

type Output

Required Methods

fn checked_shl(self, rhs: Rhs) -> Option<Self::Output>

Checked shift left. Computes self << rhs, returning None if rhs is larger than or equal to the number of bits in self.

use num_traits::CheckedShl;

let x: u16 = 0x0001;

assert_eq!(CheckedShl::checked_shl(&x, 0),  Some(0x0001));
assert_eq!(CheckedShl::checked_shl(&x, 1),  Some(0x0002));
assert_eq!(CheckedShl::checked_shl(&x, 15), Some(0x8000));
assert_eq!(CheckedShl::checked_shl(&x, 16), None);

Object Safety

This trait is not object safe.