namada_sdk::masp_primitives::num_traits::ops::overflowing
Trait OverflowingSub

pub trait OverflowingSub<Rhs = Self>: Sized {
    type Output;

    // Required method
    fn overflowing_sub(self, v: Rhs) -> (Self::Output, bool);
}

Performs substraction with a flag for overflow.

Required Associated Types

type Output

Required Methods

fn overflowing_sub(self, v: Rhs) -> (Self::Output, bool)

Returns a tuple of the difference along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

Object Safety

This trait is not object safe.