namada_sdk::masp_primitives::num_traits
Trait CheckedDiv

pub trait CheckedDiv<Rhs = Self>: Sized {
    type Output;

    // Required method
    fn checked_div(self, v: Rhs) -> Option<Self::Output>;
}

Performs division, returning None on division by zero or if overflow occurred.

Required Associated Types

type Output

Required Methods

fn checked_div(self, v: Rhs) -> Option<Self::Output>

Divides two numbers, checking for overflow and division by zero. If any of that happens, None is returned.

Object Safety

This trait is not object safe.