namada_sdk::masp_primitives::sapling::note_encryption
Function sapling_note_encryption

pub fn sapling_note_encryption<P>(
    ovk: Option<OutgoingViewingKey>,
    note: Note,
    to: PaymentAddress,
    memo: MemoBytes,
) -> NoteEncryption<SaplingDomain<P>>
where
    P: Parameters,

Creates a new encryption context for the given note.

Setting ovk to None represents the ovk = ⊥ case, where the note cannot be recovered by the sender.

NB: the example code here only covers the post-MASP case.

Examples

use ff::Field;
use rand_core::OsRng;
use masp_primitives::{
    asset_type::AssetType,
    keys::{OutgoingViewingKey, prf_expand},
    consensus::{TEST_NETWORK, TestNetwork, NetworkUpgrade, Parameters},
    memo::MemoBytes,
    sapling::{
        note_encryption::sapling_note_encryption,
        util::generate_random_rseed,
        Diversifier, PaymentAddress, Rseed, ValueCommitment
    },
};

let mut rng = OsRng;

let diversifier = Diversifier([10u8; 11]);
let pk_d = diversifier.g_d().unwrap();
let to = PaymentAddress::from_parts(diversifier, pk_d).unwrap();
let ovk = Some(OutgoingViewingKey([0; 32]));

let value = 1000;
let rcv = jubjub::Fr::random(&mut rng);
let asset_type = AssetType::new(b"note_encryption").unwrap();
let cv = asset_type.value_commitment(1, jubjub::Fr::random(&mut rng));

let height = TEST_NETWORK.activation_height(NetworkUpgrade::MASP).unwrap();
let rseed = generate_random_rseed(&TEST_NETWORK, height, &mut rng);
let note = to.create_note(asset_type, value, rseed).unwrap();
let cmu = note.cmu();

let mut enc = sapling_note_encryption::<TestNetwork>(ovk, note, to, MemoBytes::empty());
let encCiphertext = enc.encrypt_note_plaintext();
let outCiphertext = enc.encrypt_outgoing_plaintext(&cv.commitment().into(), &cmu, &mut rng);