namada_sdk::masp_primitives::sapling::note_encryption
Function try_sapling_output_recovery_with_ock

pub fn try_sapling_output_recovery_with_ock<P>(
    params: &P,
    height: BlockHeight,
    ock: &OutgoingCipherKey,
    output: &OutputDescription<[u8; 192]>,
) -> Option<(Note, PaymentAddress, MemoBytes)>
where
    P: Parameters,

Recovery of the full note plaintext by the sender.

Attempts to decrypt and validate the given enc_ciphertext using the given ock. If successful, the corresponding Sapling note and memo are returned, along with the PaymentAddress to which the note was sent.

Implements part of section 4.19.3 of the Zcash Protocol Specification. For decryption using a Full Viewing Key see try_sapling_output_recovery.