namada_sdk::masp_primitives::sapling::prover
Trait TxProver

pub trait TxProver {
    type SaplingProvingContext;

    // Required methods
    fn new_sapling_proving_context(&self) -> Self::SaplingProvingContext;
    fn spend_proof(
        &self,
        ctx: &mut Self::SaplingProvingContext,
        proof_generation_key: ProofGenerationKey,
        diversifier: Diversifier,
        rseed: Rseed,
        ar: Fr,
        asset_type: AssetType,
        value: u64,
        anchor: Scalar,
        merkle_path: MerklePath<Node>,
        rcv: Fr,
    ) -> Result<([u8; 192], ExtendedPoint, PublicKey), ()>;
    fn output_proof(
        &self,
        ctx: &mut Self::SaplingProvingContext,
        esk: Fr,
        payment_address: PaymentAddress,
        rcm: Fr,
        asset_type: AssetType,
        value: u64,
        rcv: Fr,
    ) -> ([u8; 192], ExtendedPoint);
    fn convert_proof(
        &self,
        ctx: &mut Self::SaplingProvingContext,
        allowed_conversion: AllowedConversion,
        value: u64,
        anchor: Scalar,
        merkle_path: MerklePath<Node>,
        rcv: Fr,
    ) -> Result<([u8; 192], ExtendedPoint), ()>;
    fn binding_sig(
        &self,
        ctx: &mut Self::SaplingProvingContext,
        amount: &ValueSum<AssetType, i128>,
        sighash: &[u8; 32],
    ) -> Result<Signature, ()>;
}

Interface for creating zero-knowledge proofs for shielded transactions.

Required Associated Types

type SaplingProvingContext

Type for persisting any necessary context across multiple Sapling proofs.

Required Methods

fn new_sapling_proving_context(&self) -> Self::SaplingProvingContext

Instantiate a new Sapling proving context.

fn spend_proof( &self, ctx: &mut Self::SaplingProvingContext, proof_generation_key: ProofGenerationKey, diversifier: Diversifier, rseed: Rseed, ar: Fr, asset_type: AssetType, value: u64, anchor: Scalar, merkle_path: MerklePath<Node>, rcv: Fr, ) -> Result<([u8; 192], ExtendedPoint, PublicKey), ()>

Create the value commitment, re-randomized key, and proof for a MASP SpendDescription, while accumulating its value commitment randomness inside the context for later use.

fn output_proof( &self, ctx: &mut Self::SaplingProvingContext, esk: Fr, payment_address: PaymentAddress, rcm: Fr, asset_type: AssetType, value: u64, rcv: Fr, ) -> ([u8; 192], ExtendedPoint)

Create the value commitment and proof for a MASP OutputDescription, while accumulating its value commitment randomness inside the context for later use.

fn convert_proof( &self, ctx: &mut Self::SaplingProvingContext, allowed_conversion: AllowedConversion, value: u64, anchor: Scalar, merkle_path: MerklePath<Node>, rcv: Fr, ) -> Result<([u8; 192], ExtendedPoint), ()>

Create the value commitment, and proof for a MASP ConvertDescription, while accumulating its value commitment randomness inside the context for later use.

fn binding_sig( &self, ctx: &mut Self::SaplingProvingContext, amount: &ValueSum<AssetType, i128>, sighash: &[u8; 32], ) -> Result<Signature, ()>

Create the bindingSig for a Sapling transaction. All calls to TxProver::spend_proof and TxProver::output_proof must be completed before calling this function.