namada_sdk::masp_primitives::transaction::components::sapling
Module builder

Types and functions for building MASP shielded transaction components.

Structs

ConvertBuildParams

Parameters that go into constructing an output description

ConvertDescriptionInfo

A struct containing the information required in order to construct a MASP conversion in a transaction.

OutputBuildParams

Parameters that go into constructing an output description

RngBuildParams

Lazily generated random parameters for MASP transactions

SaplingBuilder

SaplingMetadata

Metadata about a transaction created by a SaplingBuilder.

SaplingOutputInfo

A struct containing the information required in order to construct a MASP output to a transaction.

SpendBuildParams

Parameters that go into constructing a spend description

SpendDescriptionInfo

StoredBuildParams

Pre-generated random parameters for MASPtTransactions

Unauthorized

Enums

Error

Traits

BuildParams

A subset of the parameters necessary to build a transaction

MapBuilder