namada_sdk::masp_primitives::transaction::components::sapling::builder
Trait BuildParams

pub trait BuildParams {
    // Required methods
    fn spend_rcv(&mut self, i: usize) -> Fr;
    fn spend_alpha(&mut self, i: usize) -> Fr;
    fn auth_sig(&mut self, i: usize) -> Option<Signature>;
    fn proof_generation_key(&mut self, i: usize) -> Option<ProofGenerationKey>;
    fn convert_rcv(&mut self, i: usize) -> Fr;
    fn output_rcv(&mut self, i: usize) -> Fr;
    fn output_rcm(&mut self, i: usize) -> Fr;
    fn output_rseed(&mut self, i: usize) -> [u8; 32];
}

A subset of the parameters necessary to build a transaction

Required Methods

fn spend_rcv(&mut self, i: usize) -> Fr

Get the commitment value randomness for the ith spend description

fn spend_alpha(&mut self, i: usize) -> Fr

Get the spend authorization randomizer for the ith spend description

fn auth_sig(&mut self, i: usize) -> Option<Signature>

Get the authorization signature for the ith spend description

fn proof_generation_key(&mut self, i: usize) -> Option<ProofGenerationKey>

The proof generation key for the ith spend description

fn convert_rcv(&mut self, i: usize) -> Fr

Get the commitment value randomness for the ith convert description

fn output_rcv(&mut self, i: usize) -> Fr

Get the commitment value randomness for the ith output description

fn output_rcm(&mut self, i: usize) -> Fr

Get the note RCM for the ith output description

fn output_rseed(&mut self, i: usize) -> [u8; 32]

Get the random seed for the ith output description