namada_sdk::masp_primitives::transaction::components
Type Alias I8Sum

pub type I8Sum = ValueSum<AssetType, i8>;

A type-safe representation of some quantity of Zcash.

An ValueSum can only be constructed from an integer that is within the valid monetary range of {-MAX_MONEY..MAX_MONEY} (where MAX_MONEY = i64::MAX). However, this range is not preserved as an invariant internally; it is possible to add two valid ValueSums together to obtain an invalid ValueSum. It is the user’s responsibility to handle the result of serializing potentially-invalid ValueSums. In particular, a Transaction containing serialized invalid ValueSums will be rejected by the network consensus rules.

Aliased Type

struct I8Sum(/* private fields */);