namada_sdk::masp_primitives::transaction::sighash
Function signature_hash

pub fn signature_hash<TA, SA, A>(
    tx: &TransactionData<A>,
    signable_input: &SignableInput,
    txid_parts: &TxDigests<Hash>,
) -> SignatureHash
where
    TA: TransparentAuthorizingContext,
    SA: Authorization<Proof = [u8; 192]>,
    A: Authorization<SaplingAuth = SA, TransparentAuth = TA>,

Computes the signature hash for an input to a transaction, given the full data of the transaction, the input being signed, and the set of precomputed hashes produced in the construction of the transaction ID.