namada_sdk
Module queries

Ledger read-only queries can be handled and dispatched via the RPC defined via router! macro.

Modules

vp

Queries router and handlers for validity predicates

Structs

Erc20FlowControl

Contains information about the flow control of some ERC20 wrapped asset.

GenBridgePoolProofReq

Request data to pass to generate_bridge_pool_proof.

GenBridgePoolProofRsp

Response data returned by generate_bridge_pool_proof.

RequestCtx

A request context provides read-only access to storage and WASM compilation caches to request handlers.

RequestQuery

Queries for data from the application at current or past height.

ResponseQuery

Generic response from a query

Rpc

Rpcpath router type

Shell

Shellpath router type

TransferToEthereumStatus

Container for the status of queried transfers to Ethereum.

Enums

Error

Constants

RPC

RPC path router

Traits

Router

A Router handles parsing read-only query requests and dispatching them to their handler functions. A valid query returns a borsh-encoded result.

Functions

handle_path

Handle RPC query request in the ledger. On success, returns response with borsh-encoded data.

require_latest_height

For queries that only support latest height, check that the given height is not different from latest height, otherwise return an error.

require_no_data

For queries that don’t use request data, require that there are no data attached.

require_no_proof

For queries that do not support proofs, check that proof is not requested, otherwise return an error.

Type Aliases

EncodedResponseQuery

ResponseQuery with borsh-encoded data field

TransferToErcArgs

Arguments to pass to transfer_to_erc.