namada_sdk::queries
Function require_latest_height

pub fn require_latest_height<D, H, V, T>(
    ctx: &RequestCtx<'_, D, H, V, T>,
    request: &RequestQuery,
) -> Result<()>
where
    D: 'static + DB + for<'iter> DBIter<'iter> + Sync,
    H: 'static + StorageHasher + Sync,

For queries that only support latest height, check that the given height is not different from latest height, otherwise return an error.