namada_sdk::queries
Trait Router

pub trait Router {
    // Required method
    fn internal_handle<D, H, V, T>(
        &self,
        ctx: RequestCtx<'_, D, H, V, T>,
        request: &RequestQuery,
        start: usize,
    ) -> Result<EncodedResponseQuery>
       where D: 'static + DB + for<'iter> DBIter<'iter> + Sync,
             H: 'static + StorageHasher + Sync;

    // Provided method
    fn handle<D, H, V, T>(
        &self,
        ctx: RequestCtx<'_, D, H, V, T>,
        request: &RequestQuery,
    ) -> Result<EncodedResponseQuery>
       where D: 'static + DB + for<'iter> DBIter<'iter> + Sync,
             H: 'static + StorageHasher + Sync { ... }
}

A Router handles parsing read-only query requests and dispatching them to their handler functions. A valid query returns a borsh-encoded result.

Required Methods

fn internal_handle<D, H, V, T>( &self, ctx: RequestCtx<'_, D, H, V, T>, request: &RequestQuery, start: usize, ) -> Result<EncodedResponseQuery>
where D: 'static + DB + for<'iter> DBIter<'iter> + Sync, H: 'static + StorageHasher + Sync,

Internal method which shouldn’t be invoked directly. Instead, you may want to call self.handle().

Handle a given request using the provided context, starting to try to match request.path against the Router’s patterns at the given start offset.

Provided Methods

fn handle<D, H, V, T>( &self, ctx: RequestCtx<'_, D, H, V, T>, request: &RequestQuery, ) -> Result<EncodedResponseQuery>
where D: 'static + DB + for<'iter> DBIter<'iter> + Sync, H: 'static + StorageHasher + Sync,

Handle a given request using the provided context. This must be invoked on the root Router to be able to match the request.path fully.

Object Safety

This trait is not object safe.