namada_sdk::queries::vp
Module pos

Queries router and handlers for PoS validity predicate

Modules

client_only_methods

Client-only methods for the router type are composed from router functions.

Structs

Enriched

Enriched bonds data with extra information calculated from the data queried from the node.

Pos

Pospath router type

Constants

POS

POS path router

Type Aliases

EnrichedBondsAndUnbondsDetail

Bonds and unbonds with all details (slashes and rewards, if any) enriched with extra information calculated from the data queried from the node.

EnrichedBondsAndUnbondsDetails

Bonds and unbonds with all details (slashes and rewards, if any) grouped by their bond IDs enriched with extra information calculated from the data queried from the node.

ValidatorStateInfo

The validator state from a query and the epoch when it was queried.