namada_sdk::queries::vp::pos
Struct Pos

pub struct Pos { /* private fields */ }

Pospath router type

Implementations

impl Pos

pub async fn enriched_bonds_and_unbonds<CLIENT>( &self, client: &CLIENT, current_epoch: Epoch, source: &Option<Address>, validator: &Option<Address>, ) -> Result<EnrichedBondsAndUnbondsDetails, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Get bonds and unbonds with all details (slashes and rewards, if any) grouped by their bond IDs, enriched with extra information calculated from the data.

impl Pos

pub const fn new() -> Self

Construct this router as a root router

pub const fn sub(prefix: String) -> Self

Construct this router as a sub-router at the given prefix path

pub fn has_bonds_path(&self, source: &Address) -> String

Get a path to query has_bonds.

pub async fn has_bonds<CLIENT>( &self, client: &CLIENT, source: &Address, ) -> Result<bool, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from has_bonds, without any additional request data, specified block height or proof.

pub fn consensus_key_set_path(&self) -> String

Get a path to query consensus_key_set.

pub async fn consensus_key_set<CLIENT>( &self, client: &CLIENT, ) -> Result<BTreeSet<PublicKey>, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from consensus_key_set, without any additional request data, specified block height or proof.

pub fn validator_by_tm_addr_path(&self, tm_addr: &String) -> String

Get a path to query validator_by_tm_addr.

pub async fn validator_by_tm_addr<CLIENT>( &self, client: &CLIENT, tm_addr: &String, ) -> Result<Option<Address>, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from validator_by_tm_addr, without any additional request data, specified block height or proof.

pub fn is_delegator_path(&self, addr: &Address, epoch: &Option<Epoch>) -> String

Get a path to query is_delegator.

pub async fn is_delegator<CLIENT>( &self, client: &CLIENT, addr: &Address, epoch: &Option<Epoch>, ) -> Result<bool, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from is_delegator, without any additional request data, specified block height or proof.

pub fn slashes_path(&self) -> String

Get a path to query slashes.

pub async fn slashes<CLIENT>( &self, client: &CLIENT, ) -> Result<HashMap<Address, Vec<Slash>>, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from slashes, without any additional request data, specified block height or proof.

pub fn enqueued_slashes_path(&self) -> String

Get a path to query enqueued_slashes.

pub async fn enqueued_slashes<CLIENT>( &self, client: &CLIENT, ) -> Result<HashMap<Address, BTreeMap<Epoch, Vec<Slash>>>, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from enqueued_slashes, without any additional request data, specified block height or proof.

pub fn bonds_and_unbonds_path( &self, source: &Option<Address>, validator: &Option<Address>, ) -> String

Get a path to query bonds_and_unbonds.

pub async fn bonds_and_unbonds<CLIENT>( &self, client: &CLIENT, source: &Option<Address>, validator: &Option<Address>, ) -> Result<BondsAndUnbondsDetails, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from bonds_and_unbonds, without any additional request data, specified block height or proof.

pub fn withdrawable_tokens_path( &self, source: &Address, validator: &Address, epoch: &Option<Epoch>, ) -> String

Get a path to query withdrawable_tokens.

pub async fn withdrawable_tokens<CLIENT>( &self, client: &CLIENT, source: &Address, validator: &Address, epoch: &Option<Epoch>, ) -> Result<Amount, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from withdrawable_tokens, without any additional request data, specified block height or proof.

pub fn unbond_with_slashing_path( &self, source: &Address, validator: &Address, ) -> String

Get a path to query unbond_with_slashing.

pub async fn unbond_with_slashing<CLIENT>( &self, client: &CLIENT, source: &Address, validator: &Address, ) -> Result<HashMap<(Epoch, Epoch), Amount>, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from unbond_with_slashing, without any additional request data, specified block height or proof.

pub fn unbond_path(&self, source: &Address, validator: &Address) -> String

Get a path to query unbond.

pub async fn unbond<CLIENT>( &self, client: &CLIENT, source: &Address, validator: &Address, ) -> Result<HashMap<(Epoch, Epoch), Amount>, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from unbond, without any additional request data, specified block height or proof.

pub fn bond_with_slashing_path( &self, source: &Address, validator: &Address, epoch: &Option<Epoch>, ) -> String

Get a path to query bond_with_slashing.

pub async fn bond_with_slashing<CLIENT>( &self, client: &CLIENT, source: &Address, validator: &Address, epoch: &Option<Epoch>, ) -> Result<Amount, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from bond_with_slashing, without any additional request data, specified block height or proof.

pub fn rewards_path( &self, validator: &Address, source: &Option<Address>, ) -> String

Get a path to query rewards.

pub async fn rewards<CLIENT>( &self, client: &CLIENT, validator: &Address, source: &Option<Address>, ) -> Result<Amount, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from rewards, without any additional request data, specified block height or proof.

pub fn bond_path( &self, source: &Address, validator: &Address, epoch: &Option<Epoch>, ) -> String

Get a path to query bond.

pub async fn bond<CLIENT>( &self, client: &CLIENT, source: &Address, validator: &Address, epoch: &Option<Epoch>, ) -> Result<Amount, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from bond, without any additional request data, specified block height or proof.

pub fn bond_deltas_path(&self, source: &Address, validator: &Address) -> String

Get a path to query bond_deltas.

pub async fn bond_deltas<CLIENT>( &self, client: &CLIENT, source: &Address, validator: &Address, ) -> Result<HashMap<Epoch, Change>, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from bond_deltas, without any additional request data, specified block height or proof.

pub fn delegations_path(&self, owner: &Address, epoch: &Option<Epoch>) -> String

Get a path to query delegations.

pub async fn delegations<CLIENT>( &self, client: &CLIENT, owner: &Address, epoch: &Option<Epoch>, ) -> Result<HashMap<Address, Amount>, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from delegations, without any additional request data, specified block height or proof.

pub fn delegation_validators_path( &self, owner: &Address, epoch: &Option<Epoch>, ) -> String

Get a path to query delegation_validators.

pub async fn delegation_validators<CLIENT>( &self, client: &CLIENT, owner: &Address, epoch: &Option<Epoch>, ) -> Result<HashSet<Address>, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from delegation_validators, without any additional request data, specified block height or proof.

pub fn total_active_voting_power_path(&self, epoch: &Option<Epoch>) -> String

Get a path to query total_active_voting_power.

pub async fn total_active_voting_power<CLIENT>( &self, client: &CLIENT, epoch: &Option<Epoch>, ) -> Result<Amount, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from total_active_voting_power, without any additional request data, specified block height or proof.

pub fn total_stake_path(&self, epoch: &Option<Epoch>) -> String

Get a path to query total_stake.

pub async fn total_stake<CLIENT>( &self, client: &CLIENT, epoch: &Option<Epoch>, ) -> Result<Amount, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from total_stake, without any additional request data, specified block height or proof.

pub fn pos_params_path(&self) -> String

Get a path to query pos_params.

pub async fn pos_params<CLIENT>( &self, client: &CLIENT, ) -> Result<PosParams, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from pos_params, without any additional request data, specified block height or proof.

pub fn consensus_validator_set_path(&self, epoch: &Option<Epoch>) -> String

Get a path to query consensus_validator_set.

pub async fn consensus_validator_set<CLIENT>( &self, client: &CLIENT, epoch: &Option<Epoch>, ) -> Result<BTreeSet<WeightedValidator>, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from consensus_validator_set, without any additional request data, specified block height or proof.

pub fn below_capacity_validator_set_path(&self, epoch: &Option<Epoch>) -> String

Get a path to query below_capacity_validator_set.

pub async fn below_capacity_validator_set<CLIENT>( &self, client: &CLIENT, epoch: &Option<Epoch>, ) -> Result<BTreeSet<WeightedValidator>, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from below_capacity_validator_set, without any additional request data, specified block height or proof.

pub fn is_validator_path(&self, addr: &Address) -> String

Get a path to query is_validator.

pub async fn is_validator<CLIENT>( &self, client: &CLIENT, addr: &Address, ) -> Result<bool, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from is_validator, without any additional request data, specified block height or proof.

pub fn consensus_key_path(&self, addr: &Address) -> String

Get a path to query consensus_key.

pub async fn consensus_key<CLIENT>( &self, client: &CLIENT, addr: &Address, ) -> Result<Option<PublicKey>, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from consensus_key, without any additional request data, specified block height or proof.

pub fn validator_addresses_path(&self, epoch: &Option<Epoch>) -> String

Get a path to query validator_addresses.

pub async fn validator_addresses<CLIENT>( &self, client: &CLIENT, epoch: &Option<Epoch>, ) -> Result<HashSet<Address>, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from validator_addresses, without any additional request data, specified block height or proof.

pub fn liveness_info_path(&self) -> String

Get a path to query liveness_info.

pub async fn liveness_info<CLIENT>( &self, client: &CLIENT, ) -> Result<LivenessInfo, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from liveness_info, without any additional request data, specified block height or proof.

pub fn validator_stake_path( &self, validator: &Address, epoch: &Option<Epoch>, ) -> String

Get a path to query validator_stake.

pub async fn validator_stake<CLIENT>( &self, client: &CLIENT, validator: &Address, epoch: &Option<Epoch>, ) -> Result<Option<Amount>, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from validator_stake, without any additional request data, specified block height or proof.

pub fn validator_slashes_path(&self, validator: &Address) -> String

Get a path to query validator_slashes.

pub async fn validator_slashes<CLIENT>( &self, client: &CLIENT, validator: &Address, ) -> Result<Vec<Slash>, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from validator_slashes, without any additional request data, specified block height or proof.

pub fn validator_commission_path( &self, validator: &Address, epoch: &Option<Epoch>, ) -> String

Get a path to query validator_commission.

pub async fn validator_commission<CLIENT>( &self, client: &CLIENT, validator: &Address, epoch: &Option<Epoch>, ) -> Result<CommissionPair, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from validator_commission, without any additional request data, specified block height or proof.

pub fn validator_metadata_path(&self, validator: &Address) -> String

Get a path to query validator_metadata.

pub async fn validator_metadata<CLIENT>( &self, client: &CLIENT, validator: &Address, ) -> Result<Option<ValidatorMetaData>, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from validator_metadata, without any additional request data, specified block height or proof.

pub fn validator_state_path( &self, validator: &Address, epoch: &Option<Epoch>, ) -> String

Get a path to query validator_state.

pub async fn validator_state<CLIENT>( &self, client: &CLIENT, validator: &Address, epoch: &Option<Epoch>, ) -> Result<ValidatorStateInfo, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from validator_state, without any additional request data, specified block height or proof.

pub fn validator_incoming_redelegation_path( &self, src_validator: &Address, delegator: &Address, ) -> String

Get a path to query validator_incoming_redelegation.

pub async fn validator_incoming_redelegation<CLIENT>( &self, client: &CLIENT, src_validator: &Address, delegator: &Address, ) -> Result<Option<Epoch>, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from validator_incoming_redelegation, without any additional request data, specified block height or proof.

pub fn validator_last_infraction_epoch_path( &self, validator: &Address, ) -> String

Get a path to query validator_last_infraction_epoch.

pub async fn validator_last_infraction_epoch<CLIENT>( &self, client: &CLIENT, validator: &Address, ) -> Result<Option<Epoch>, <CLIENT as Client>::Error>
where CLIENT: Client + Sync,

Request a simple borsh-encoded value from validator_last_infraction_epoch, without any additional request data, specified block height or proof.