namada_sdk
Module rpc

SDK RPC queries

Structs

TxResponse

A parsed event from tendermint relating to a transaction

Enums

InnerTxResult

Determines a result of an inner tx from namada_tx::data::BatchedTxResult.

TxBroadcastData

Data needed for broadcasting a tx and monitoring its progress on chain.

TxEventQuery

Represents a query for an event pertaining to the specified transaction

Functions

bonds_and_unbonds

Get bonds and unbonds with all details (slashes and rewards, if any) grouped by their bond IDs.

denominate_amount

Look up the denomination of a token in order to make a correctly denominated amount.

dry_run_tx

Dry run a transaction

enriched_bonds_and_unbonds

Get bonds and unbonds with all details (slashes and rewards, if any) grouped by their bond IDs, enriched with extra information calculated from the data.

format_denominated_amount

Look up the denomination of a token in order to format it correctly as a string.

get_account_info

Query the accunt substorage space of an address

get_all_consensus_validators

Get all consensus validators in the given epoch

get_all_validators

Get all validators in the given epoch

get_bond_amount_at

Get the bond amount at the given epoch

get_consensus_keys

Get the set of consensus keys registered in the network

get_delegation_validators

Get the validators to which a delegator is bonded at a certain epoch

get_delegations_of_delegator_at

Get the delegations of a delegator at some epoch, including the validator and bond amount

get_effective_native_supply

Query the effective total supply of the native token

get_pos_params

Get the PoS parameters

get_public_key_at

Query an account substorage at a specific index

get_registry_from_xcs_osmosis_contract

Query the registry contract embedded in the state of an input Crosschain Swaps Osmosis contract.

get_staking_rewards_rate

Query the effective total supply of the native token

get_token_balance

Query token amount of owner.

get_token_total_supply

Query token total supply.

get_total_active_voting_power

Get the total active voting power in the given epoch

get_total_staked_tokens

Get the total staked tokens in the given epoch

get_validator_stake

Get the given validator’s stake at the given epoch

get_validator_state

Query and return a validator’s state

get_validators_liveness_info

Get liveness info for all consensus validators in the current epoch

has_bonds

Find if the given source address has any bonds.

is_delegator

Check if a given address is a known delegator

is_delegator_at

Check if a given address is a known delegator at the given epoch

is_public_key_revealed

Query if the public_key is revealed

is_steward

Check if the given address is a pgf steward.

is_validator

Check if the given address is a known validator.

known_address

Check if the address exists on chain. Established address exists if it has a stored validity predicate. Implicit and internal addresses always return true.

osmosis_denom_from_namada_denom

Given a Namada asset returned from an Osmosis swap, find the corresponding asset denom on Osmosis.

query_and_print_unbonds

Query a validator’s unbonds for a given epoch

query_block

Query the last committed block, if any.

query_block_header

Query the epoch of the last committed block

query_bond

Query a validator’s bonds for a given epoch

query_commission_rate

Query and return validator’s commission rate and max commission rate change per epoch

query_conversion

Query a conversion.

query_conversions

Query conversions

query_denom

Query the denomination of the given token

query_epoch

Query the epoch of the last committed block

query_epoch_at_height

Query the epoch of the given block height, if it exists. Will return none if the input block height is greater than the latest committed block height.

query_governance_parameters

Get the governance parameters

query_has_storage_key

Query to check if the given storage key exists.

query_ibc_denom

Look up the IBC denomination from a IbcToken.

query_ibc_params

Query the global IBC parameters

query_ibc_rate_limits

Query the IBC rate limit for the provided token

query_ibc_tokens

Look up IBC tokens. The given base token can be non-Namada token.

query_incoming_redelegations

Query and return the incoming redelegation epoch for a given pair of source validator and delegator, if there is any.

query_last_infraction_epoch

Query a validator’s bonds for a given epoch

query_masp_epoch

Query the masp epoch of the last committed block

query_masp_reward_tokens

Query to read the tokens that earn masp rewards.

query_masp_total_rewards

Query the total rewards minted by MASP

query_max_block_time_estimate

Query an estimate of the maximum block time.

query_metadata

Query and return validator’s metadata, including the commission rate and max commission rate change

query_native_token

Query the address of the native token

query_next_epoch_info

Query the information to estimate next epoch start

query_osmosis_pool_routes

Query a route of Osmosis liquidity pools for swapping betwixt token and output_denom assets.

query_pgf_fundings

Get the set of pgf fundings

query_pgf_parameters

Get the public good fundings parameters

query_pgf_stewards

Get the set of pgf stewards

query_proposal_by_id

Query proposal by Id

query_proposal_result

Query the proposal result

query_proposal_votes

Get all the votes of a proposal

query_results

Query the results of the last committed block

query_rewards

Query and return the available reward tokens corresponding to the bond

query_storage_prefix

Query a range of storage values with a matching prefix and decode them with BorshDeserialize. Returns an iterator of the storage keys paired with their associated values.

query_storage_value

Query a storage value and decode it with BorshDeserialize.

query_storage_value_bytes

Query a storage value and the proof without decoding.

query_tx_events

Call the corresponding tx_event_query RPC method, to fetch the current status of a transaction.

query_tx_status

Query the status of a given transaction.

query_tx_status2

Identical to query_tx_status, but does not need a Namada context.

query_unbond_with_slashing

Query all unbonds for a validator, applying slashes

query_validator_consensus_keys

Query the consensus key by validator address

query_wasm_code_hash

Query a wasm code hash

query_withdrawable_tokens

Query withdrawable tokens in a validator account for a given epoch

validate_amount

Get the correct representation of the amount given the token type.

wait_until_node_is_synched

Wait for a first block and node to be synced.