namada_sdk
Module signing

Functions to sign transactions

Structs

LedgerVector

Represents the transaction data that is displayed on a Ledger device

OfflineSignatures

The transaction’s signatures produced via the offline signing procedure

SigningTxData

A structure holding the signing data to craft a transaction

TxSourcePostBalance

Information about the post-fee balance of the tx’s source. Used to correctly handle balance validation in the inner tx

Enums

AddressVpType

Grouping of addresses by validity predicate.

Signable

The different parts of a transaction that can be signed. Note that it’s impossible to sign the fee header without signing the raw header.

Functions

aux_signing_data

Return the necessary data regarding an account to be able to generate a signature section

default_sign

Causes sign_tx to attempt signing using only the software wallet

find_key_by_pk

Load the secret key corresponding to the given public key from the wallet.

find_pk

Find the public key for the given address and try to load the keypair for it from the wallet.

gen_disposable_signing_key

Generate a disposable signing key.

generate_tx_signatures

Generates the transaction’s signatures for offline signing purposes. This allows to sign both the inner transactions of the batch as well as the wrapper transaction and it supports multisignatures.

sign_tx

Sign a transaction with a given signing key or public key of a given signer. If no explicit signer given, use the default. If no default is given, Error.

to_ledger_vector

Converts the given transaction to the form that is displayed on the Ledger device

tx_signers

Given CLI arguments and some defaults, determine the rightful transaction signer.

validate_fee

Validate the fee amount and token

validate_transparent_fee

Validate the fee of the transaction in case of a transparent fee payer, computing the updated post balance

wrap_tx

Create a wrapper tx from a normal tx. Get the hash of the wrapper and its payload which is needed for monitoring its progress on chain.