namada_sdk
Struct NamadaImpl

pub struct NamadaImpl<C, U, V, I>
where
    C: Client,
    U: WalletIo,
    V: ShieldedUtils,
    I: Io,
{
    pub client: C,
    pub wallet: RwLock<Wallet<U>>,
    pub shielded: RwLock<ShieldedContext<V>>,
    pub io: I,
    /* private fields */
}

Provides convenience methods for common Namada interactions

Fields
client: C

Used to send and receive messages from the ledger

wallet: RwLock<Wallet<U>>

Stores the addresses and keys required for ledger interactions

shielded: RwLock<ShieldedContext<V>>

Stores the current state of the shielded pool

io: I

Captures the input/output streams used by this object

Implementations

impl<C, U, V, I> NamadaImpl<C, U, V, I>
where C: Client + Sync, U: WalletIo, V: ShieldedUtils, I: Io,

pub fn native_new( client: C, wallet: Wallet<U>, shielded: ShieldedWallet<V>, io: I, native_token: Address, ) -> Self

Construct a new Namada context with the given native token address

pub async fn new( client: C, wallet: Wallet<U>, shielded: ShieldedWallet<V>, io: I, ) -> Result<NamadaImpl<C, U, V, I>>

Construct a new Namada context looking up the native token address

pub fn clone_client(&self) -> C
where C: Clone,

Clone the rpc client.