namada_sdk::tendermint::crypto
Module default

Pure Rust implementations of the cryptographic traits.

Most applications using this crate should use these implementations. Alternative implementations can be useful on targets like wasm and on-chain environments, where code size is at a premium and a faster platform-native cryptographic API is available.

Modules

ecdsa_secp256k1

Types implementing the ECDSA algorithm using the Secp256k1 elliptic curve.

signature

The pure Rust implementation of signature verification functions.

Type Aliases

Sha256

SHA-256 hasher.