namada_sdk::tendermint::merkle
Function simple_hash_from_byte_vectors

pub fn simple_hash_from_byte_vectors<H>(
    byte_vecs: &[impl AsRef<[u8]>],
) -> [u8; 32]
where
    H: MerkleHash + Default,

Compute a simple Merkle root from vectors of arbitrary byte vectors. The leaves of the tree are the bytes of the given byte vectors in the given order.