namada_sdk::tendermint::serializers
Module allow_null

Serialize/deserialize nilable value into T, where nil turns into the Default value.

Serialize any value of T, including the default value, using the serialization for Some variant of Option<T>.

This helper can be used to tolerate nil values from a serialization producer, while the default value is normatively serialized as such.

Functions

deserialize

Deserialize T from a nil-able representation, accepting the nil

serialize

Serialize T as Some value of Option<T>.