namada_sdk::tendermint::trust_threshold
Trait TrustThreshold

pub trait TrustThreshold:
    Copy
    + Clone
    + Debug
    + Serialize
    + DeserializeOwned {
    // Required method
    fn is_enough_power(
        &self,
        signed_voting_power: u64,
        total_voting_power: u64,
    ) -> bool;
}

TrustThreshold defines how much of the total voting power of a known and trusted validator set is sufficient for a commit to be accepted going forward.

Required Methods

fn is_enough_power( &self, signed_voting_power: u64, total_voting_power: u64, ) -> bool

Check whether the given signed voting power is sufficient according to this trust threshold against the given total voting power.

Object Safety

This trait is not object safe.