namada_sdk::tendermint::v0_38::abci
Module response

Structs

ApplySnapshotChunk

Returns the result of applying a snapshot chunk and associated data.

BeginBlock

Returns events that occurred when beginning a new block.

CheckTx

Returns the result of checking a transaction for mempool inclusion.

Commit

Returns the result of persisting the application state.

DeliverTx

Returns events that occurred while executing a transaction against the application state.

Echo

Echoes a string to test an ABCI implementation.

EndBlock

Returns validator updates that occur after the end of a block.

Exception

Returns an exception (undocumented, nondeterministic).

ExtendVote

ABCI documentation

FinalizeBlock

Info

Returns information about the application state.

InitChain

Returned on genesis after initializing chain state.

ListSnapshots

Returns a list of local state snapshots.

LoadSnapshotChunk

Returns a snapshot chunk from the application.

PrepareProposal

ABCI documentation

Query

Returns data queried from the application.

Enums

ConsensusResponse

The consensus category of ABCI responses.

InfoResponse

The info category of ABCI responses.

MempoolResponse

The mempool category of ABCI responses.

OfferSnapshot

Returns the application’s response to a snapshot offer.

ProcessProposal

ABCI documentation

Response

All possible ABCI responses for this protocol version.

SnapshotResponse

The snapshot category of ABCI responses.

VerifyVoteExtension

ABCI documentation