namada_sdk
Trait Namada

pub trait Namada: NamadaIo {
    type WalletUtils: WalletIo + WalletStorage + MaybeSend + MaybeSync;
    type ShieldedUtils: ShieldedUtils + MaybeSend + MaybeSync;

    // Required methods
    fn wallet<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = RwLockReadGuard<'_, Wallet<Self::WalletUtils>>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn wallet_mut<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = RwLockWriteGuard<'_, Wallet<Self::WalletUtils>>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn wallet_lock(&self) -> &RwLock<Wallet<Self::WalletUtils>>;
    fn shielded<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = RwLockReadGuard<'_, ShieldedContext<Self::ShieldedUtils>>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn shielded_mut<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = RwLockWriteGuard<'_, ShieldedContext<Self::ShieldedUtils>>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn native_token(&self) -> Address;

    // Provided methods
    fn tx_builder(&self) -> Tx { ... }
    fn new_transparent_transfer(
        &self,
        data: Vec<TxTransparentTransferData>,
    ) -> TxTransparentTransfer { ... }
    fn new_shielded_transfer(
        &self,
        data: Vec<TxShieldedTransferData>,
        gas_spending_key: Option<ExtendedSpendingKey>,
        disposable_signing_key: bool,
    ) -> TxShieldedTransfer { ... }
    fn new_shielding_transfer(
        &self,
        target: PaymentAddress,
        data: Vec<TxShieldingTransferData>,
    ) -> TxShieldingTransfer { ... }
    fn new_unshielding_transfer(
        &self,
        source: ExtendedSpendingKey,
        data: Vec<TxUnshieldingTransferData>,
        gas_spending_key: Option<ExtendedSpendingKey>,
        disposable_signing_key: bool,
    ) -> TxUnshieldingTransfer { ... }
    fn new_init_account(
        &self,
        public_keys: Vec<PublicKey>,
        threshold: Option<u8>,
    ) -> TxInitAccount { ... }
    fn new_reveal_pk(&self, public_key: PublicKey) -> RevealPk { ... }
    fn new_bond(&self, validator: Address, amount: Amount) -> Bond { ... }
    fn new_unbond(&self, validator: Address, amount: Amount) -> Unbond { ... }
    fn new_redelegation(
        &self,
        source: Address,
        src_validator: Address,
        dest_validator: Address,
        amount: Amount,
    ) -> Redelegate { ... }
    fn new_ibc_transfer(
        &self,
        source: TransferSource,
        receiver: String,
        token: Address,
        amount: InputAmount,
        channel_id: ChannelId,
        disposable_signing_key: bool,
    ) -> TxIbcTransfer { ... }
    fn new_init_proposal(&self, proposal_data: Vec<u8>) -> InitProposal { ... }
    fn new_update_account(
        &self,
        addr: Address,
        public_keys: Vec<PublicKey>,
        threshold: u8,
    ) -> TxUpdateAccount { ... }
    fn new_proposal_vote(
        &self,
        proposal_id: u64,
        vote: String,
        voter_address: Address,
    ) -> VoteProposal { ... }
    fn new_change_commission_rate(
        &self,
        rate: Dec,
        validator: Address,
    ) -> CommissionRateChange { ... }
    fn new_change_consensus_key(
        &self,
        validator: Address,
        consensus_key: PublicKey,
    ) -> ConsensusKeyChange { ... }
    fn new_change_metadata(&self, validator: Address) -> MetaDataChange { ... }
    fn new_become_validator(
        &self,
        address: Address,
        commission_rate: Dec,
        max_commission_rate_change: Dec,
        consesus_key: PublicKey,
        eth_cold_key: PublicKey,
        eth_hot_key: PublicKey,
        protocol_key: PublicKey,
        email: String,
    ) -> TxBecomeValidator { ... }
    fn new_init_validator(
        &self,
        commission_rate: Dec,
        max_commission_rate_change: Dec,
        email: String,
    ) -> TxInitValidator { ... }
    fn new_unjail_validator(&self, validator: Address) -> TxUnjailValidator { ... }
    fn new_deactivate_validator(
        &self,
        validator: Address,
    ) -> TxDeactivateValidator { ... }
    fn new_reactivate_validator(
        &self,
        validator: Address,
    ) -> TxReactivateValidator { ... }
    fn new_withdraw(&self, validator: Address) -> Withdraw { ... }
    fn new_claim_rewards(&self, validator: Address) -> ClaimRewards { ... }
    fn new_add_erc20_transfer(
        &self,
        sender: Address,
        recipient: EthAddress,
        asset: EthAddress,
        amount: InputAmount,
    ) -> EthereumBridgePool { ... }
    fn new_resign_steward(&self, steward: Address) -> ResignSteward { ... }
    fn new_update_steward_rewards(
        &self,
        steward: Address,
        commission: Vec<u8>,
    ) -> UpdateStewardCommission { ... }
    fn new_custom(&self, owner: Option<Address>) -> TxCustom { ... }
    fn sign<'life0, 'life1, 'life2, 'async_trait, D, F>(
        &'life0 self,
        tx: &'life1 mut Tx,
        args: &'life2 Tx,
        signing_data: SigningTxData,
        with: impl 'async_trait + Fn(Tx, PublicKey, HashSet<Signable>, D) -> F + MaybeSend + MaybeSync,
        user_data: D,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + 'async_trait>>
       where D: Clone + MaybeSend + MaybeSync + 'async_trait,
             F: MaybeSend + MaybeSync + Future<Output = Result<Tx>> + 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn submit<'life0, 'life1, 'async_trait>(
        &'life0 self,
        tx: Tx,
        args: &'life1 Tx,
    ) -> Pin<Box<dyn Future<Output = Result<ProcessTxResponse>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn denominate_amount<'life0, 'life1, 'async_trait>(
        &'life0 self,
        token: &'life1 Address,
        amount: Amount,
    ) -> Pin<Box<dyn Future<Output = DenominatedAmount> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn format_amount<'life0, 'life1, 'async_trait>(
        &'life0 self,
        token: &'life1 Address,
        amount: Amount,
    ) -> Pin<Box<dyn Future<Output = String> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }

}

An interface for high-level interaction with the Namada SDK

Required Associated Types

type WalletUtils: WalletIo + WalletStorage + MaybeSend + MaybeSync

Captures the interactive parts of the wallet’s functioning

type ShieldedUtils: ShieldedUtils + MaybeSend + MaybeSync

Abstracts platform specific details away from the logic of shielded pool operations.

Required Methods

fn wallet<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = RwLockReadGuard<'_, Wallet<Self::WalletUtils>>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Obtain read guard on the wallet

fn wallet_mut<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = RwLockWriteGuard<'_, Wallet<Self::WalletUtils>>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Obtain write guard on the wallet

fn wallet_lock(&self) -> &RwLock<Wallet<Self::WalletUtils>>

Obtain the wallet lock

fn shielded<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = RwLockReadGuard<'_, ShieldedContext<Self::ShieldedUtils>>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Obtain read guard on the shielded context

fn shielded_mut<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = RwLockWriteGuard<'_, ShieldedContext<Self::ShieldedUtils>>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Obtain write guard on the shielded context

fn native_token(&self) -> Address

Return the native token

Provided Methods

fn tx_builder(&self) -> Tx

Make a tx builder using no arguments

fn new_transparent_transfer( &self, data: Vec<TxTransparentTransferData>, ) -> TxTransparentTransfer

Make a TxTransparentTransfer builder from the given minimum set of arguments

fn new_shielded_transfer( &self, data: Vec<TxShieldedTransferData>, gas_spending_key: Option<ExtendedSpendingKey>, disposable_signing_key: bool, ) -> TxShieldedTransfer

Make a TxShieldedTransfer builder from the given minimum set of arguments

fn new_shielding_transfer( &self, target: PaymentAddress, data: Vec<TxShieldingTransferData>, ) -> TxShieldingTransfer

Make a TxShieldingTransfer builder from the given minimum set of arguments

fn new_unshielding_transfer( &self, source: ExtendedSpendingKey, data: Vec<TxUnshieldingTransferData>, gas_spending_key: Option<ExtendedSpendingKey>, disposable_signing_key: bool, ) -> TxUnshieldingTransfer

Make a TxUnshieldingTransfer builder from the given minimum set of arguments

fn new_init_account( &self, public_keys: Vec<PublicKey>, threshold: Option<u8>, ) -> TxInitAccount

Make a InitAccount builder from the given minimum set of arguments

fn new_reveal_pk(&self, public_key: PublicKey) -> RevealPk

Make a RevealPK builder from the given minimum set of arguments

fn new_bond(&self, validator: Address, amount: Amount) -> Bond

Make a Bond builder from the given minimum set of arguments

fn new_unbond(&self, validator: Address, amount: Amount) -> Unbond

Make a Unbond builder from the given minimum set of arguments

fn new_redelegation( &self, source: Address, src_validator: Address, dest_validator: Address, amount: Amount, ) -> Redelegate

Make a Redelegation builder for the given minimum set of arguments

fn new_ibc_transfer( &self, source: TransferSource, receiver: String, token: Address, amount: InputAmount, channel_id: ChannelId, disposable_signing_key: bool, ) -> TxIbcTransfer

Make a TxIbcTransfer builder from the given minimum set of arguments

fn new_init_proposal(&self, proposal_data: Vec<u8>) -> InitProposal

Make a InitProposal builder from the given minimum set of arguments

fn new_update_account( &self, addr: Address, public_keys: Vec<PublicKey>, threshold: u8, ) -> TxUpdateAccount

Make a TxUpdateAccount builder from the given minimum set of arguments

fn new_proposal_vote( &self, proposal_id: u64, vote: String, voter_address: Address, ) -> VoteProposal

Make a VoteProposal builder from the given minimum set of arguments

fn new_change_commission_rate( &self, rate: Dec, validator: Address, ) -> CommissionRateChange

Make a CommissionRateChange builder from the given minimum set of arguments

fn new_change_consensus_key( &self, validator: Address, consensus_key: PublicKey, ) -> ConsensusKeyChange

Make ConsensusKeyChange builder from the given minimum set of arguments

fn new_change_metadata(&self, validator: Address) -> MetaDataChange

Make a CommissionRateChange builder from the given minimum set of arguments

fn new_become_validator( &self, address: Address, commission_rate: Dec, max_commission_rate_change: Dec, consesus_key: PublicKey, eth_cold_key: PublicKey, eth_hot_key: PublicKey, protocol_key: PublicKey, email: String, ) -> TxBecomeValidator

Make a TxBecomeValidator builder from the given minimum set of arguments

fn new_init_validator( &self, commission_rate: Dec, max_commission_rate_change: Dec, email: String, ) -> TxInitValidator

Make a TxInitValidator builder from the given minimum set of arguments

fn new_unjail_validator(&self, validator: Address) -> TxUnjailValidator

Make a TxUnjailValidator builder from the given minimum set of arguments

fn new_deactivate_validator(&self, validator: Address) -> TxDeactivateValidator

Make a TxDeactivateValidator builder from the given minimum set of arguments

fn new_reactivate_validator(&self, validator: Address) -> TxReactivateValidator

Make a TxReactivateValidator builder from the given minimum set of arguments

fn new_withdraw(&self, validator: Address) -> Withdraw

Make a Withdraw builder from the given minimum set of arguments

fn new_claim_rewards(&self, validator: Address) -> ClaimRewards

Make a Claim-rewards builder from the given minimum set of arguments

fn new_add_erc20_transfer( &self, sender: Address, recipient: EthAddress, asset: EthAddress, amount: InputAmount, ) -> EthereumBridgePool

Make a Withdraw builder from the given minimum set of arguments

fn new_resign_steward(&self, steward: Address) -> ResignSteward

Make a ResignSteward builder from the given minimum set of arguments

fn new_update_steward_rewards( &self, steward: Address, commission: Vec<u8>, ) -> UpdateStewardCommission

Make a UpdateStewardCommission builder from the given minimum set of arguments

fn new_custom(&self, owner: Option<Address>) -> TxCustom

Make a TxCustom builder from the given minimum set of arguments

fn sign<'life0, 'life1, 'life2, 'async_trait, D, F>( &'life0 self, tx: &'life1 mut Tx, args: &'life2 Tx, signing_data: SigningTxData, with: impl 'async_trait + Fn(Tx, PublicKey, HashSet<Signable>, D) -> F + MaybeSend + MaybeSync, user_data: D, ) -> Pin<Box<dyn Future<Output = Result<()>> + 'async_trait>>
where D: Clone + MaybeSend + MaybeSync + 'async_trait, F: MaybeSend + MaybeSync + Future<Output = Result<Tx>> + 'async_trait, Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Sign the given transaction using the given signing data

fn submit<'life0, 'life1, 'async_trait>( &'life0 self, tx: Tx, args: &'life1 Tx, ) -> Pin<Box<dyn Future<Output = Result<ProcessTxResponse>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Process the given transaction using the given flags

fn denominate_amount<'life0, 'life1, 'async_trait>( &'life0 self, token: &'life1 Address, amount: Amount, ) -> Pin<Box<dyn Future<Output = DenominatedAmount> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Look up the denomination of a token in order to make a correctly denominated amount.

fn format_amount<'life0, 'life1, 'async_trait>( &'life0 self, token: &'life1 Address, amount: Amount, ) -> Pin<Box<dyn Future<Output = String> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Look up the denomination of a token in order to format it correctly as a string.

Object Safety

This trait is not object safe.