namada_sdk
Module tx

SDK functions to construct different types of transactions

Modules

action

Actions applied in a transaction.

data

Data-Types that are used in transactions.

either

The enum Either with variants Left and Right is a general purpose sum type with two cases.

event

Transaction events.

proto

Generated protobuf tx type.

Structs

Authorization

A section representing a multisig over another section

BatchedTx

A transaction with the commitment to a specific inner transaction of the batch

BatchedTxRef

A reference to a transaction with the commitment to a specific inner transaction of the batch

Code

A section representing transaction code

CompressedAuthorization

A section representing a multisig over another section

Data

A section representing transaction data

Header

A Namada transaction header indicating where transaction subcomponents can be found

IndexedTx

Represents the pointers to a indexed tx, which are the block height and the index inside that block. Optionally points to a specific inner tx inside a batch if such level of granularity is required.

IndexedTxRange

Inclusive range of IndexedTx entries.

MaspBuilder

A section providing the auxiliary inputs used to construct a MASP transaction

SignableEthMessage

Tag type that indicates we should use ABI serialization to sign data in a Signable wrapper.

SignatureIndex

Signature index within a multisig

Signed

A generic signed data wrapper for serialize-able types.

Tx

A Namada transaction is represented as a header followed by a series of sections providing additional details.

TxCommitments

An inner transaction of the batch, represented by its commitments to the Code, Data and Memo sections

Enums

Commitment

Represents either some code bytes or their SHA-256 hash

DecodeError

ProcessTxResponse

Capture the result of running a transaction

Section

A section of a transaction. Carries an independent piece of information necessary for the processing of a transaction.

Signer

Indicates the list of public keys against which signatures will be verified

TxError

Constants

SALT_LENGTH

Length of the transaction sections salt

TX_BECOME_VALIDATOR_WASM

Become validator transaction WASM path

TX_BOND_WASM

Bond WASM path

TX_BRIDGE_POOL_WASM

Bridge pool WASM path

TX_CHANGE_COMMISSION_WASM

Change commission WASM path

TX_CHANGE_CONSENSUS_KEY_WASM

Change consensus key WASM path

TX_CHANGE_METADATA_WASM

Change validator metadata WASM path

TX_CLAIM_REWARDS_WASM

Claim-rewards WASM path

TX_DEACTIVATE_VALIDATOR_WASM

Deactivate validator transaction WASM path

TX_IBC_WASM

IBC transaction WASM path

TX_INIT_ACCOUNT_WASM

Initialize account transaction WASM

TX_INIT_PROPOSAL

Initialize proposal transaction WASM path

TX_REACTIVATE_VALIDATOR_WASM

Reactivate validator transaction WASM path

TX_REDELEGATE_WASM

Redelegate transaction WASM path

TX_RESIGN_STEWARD

Resign steward WASM path

TX_REVEAL_PK

Reveal public key transaction WASM path

TX_TRANSFER_WASM

Transparent transfer transaction WASM path

TX_UNBOND_WASM

Unbond WASM path

TX_UNJAIL_VALIDATOR_WASM

Unjail validator transaction WASM path

TX_UPDATE_ACCOUNT_WASM

Update validity predicate WASM path

TX_UPDATE_STEWARD_COMMISSION

Update steward commission WASM path

TX_VOTE_PROPOSAL

Vote transaction WASM path

TX_WITHDRAW_WASM

Withdraw WASM path

VP_USER_WASM

User validity predicate WASM path

Functions

broadcast_tx

Broadcast a transaction to be included in the blockchain and checks that the tx has been successfully included into the mempool of a node

build_batch

Constructs the batched tx from the provided list. Returns also the data for signing

build_become_validator

Build a pgf funding proposal governance

build_bond

Submit a transaction to bond

build_change_consensus_key

Submit validator commission rate change

build_claim_rewards

Submit transaction to withdraw an unbond

build_custom

Submit a custom transaction

build_deactivate_validator

Submit transaction to deactivate a validator

build_default_proposal

Build a default proposal governance

build_ibc_transfer

Submit an IBC transfer

build_init_account

Submit a transaction to initialize an account

build_pgf_funding_proposal

Build a pgf funding proposal governance

build_pgf_stewards_proposal

Build a pgf funding proposal governance

build_reactivate_validator

Submit transaction to deactivate a validator

build_redelegation

Redelegate bonded tokens from one validator to another

build_resign_steward

Craft transaction to resign as a steward

build_reveal_pk

Submit transaction to reveal the given public key

build_shielded_transfer

Build a shielded transfer

build_shielding_transfer

Build a shielding transfer

build_transparent_transfer

Build a transparent transfer

build_unbond

Submit a transaction to unbond

build_unjail_validator

Submit transaction to unjail a jailed validator

build_unshielding_transfer

Build an unshielding transfer

build_update_account

Submit a transaction to update a VP

build_update_steward_commission

Craft transaction to update a steward commission

build_validator_commission_change

Submit validator commission rate change

build_validator_metadata_change

Submit validator metadata change

build_vote_proposal

Build a proposal vote

build_withdraw

Submit transaction to withdraw an unbond

display_batch_resp

Display a result of a tx batch.

dump_tx

Build and dump a transaction either to file or to screen

gen_ibc_shielding_transfer

Generate IBC shielded transfer

has_revealed_pk

Check if the public key for the given address has been revealed

is_reveal_pk_needed

Check if a reveal public key transaction is needed

new_tx_event

Creates a new event with the hash and height of the transaction already filled in.

prepare_tx

Prepare a transaction for signing and submission by adding a wrapper header to it.

process_tx

Submit transaction and wait for result. Returns a list of addresses initialized in the transaction if any. In dry run, this is always empty.

query_unbonds

Query the unbonds post-tx

save_initialized_accounts

Save accounts initialized from a tx into the wallet, if any.

standalone_signature

Get a signature for data

submit_tx

Broadcast a transaction to be included in the blockchain.

verify_standalone_sig

Verify that the input data has been signed by the secret key counterpart of the given public key.

Type Aliases

Memo

A memo field (bytes).