namada_sdk::tx
Module action

Actions applied in a transaction.

Tx actions are used to indicate from tx to VPs the type of actions that have been applied by the tx to simplify validation (We can check that the storage changes are valid based on the action, rather than trying to derive the action from storage changes). When used, the kind is expected to written to under temporary storage (discarded after tx execution and validation).

Structs

Bond

A bond is a validator’s self-bond or a delegation from non-validator to a validator.

ClaimRewards

A claim of pending rewards.

Redelegation

A redelegation of bonded tokens from one validator to another.

Withdraw

A withdrawal of an unbond.

Enums

Action

An action applied from a tx.

GovAction

Gov tx actions.

MaspAction

MASP tx actions.

PgfAction

PGF tx actions.

PosAction

PoS tx actions.

Traits

Read

Read actions from temporary storage

Write

Write actions to temporary storage

Functions

get_masp_section_ref

Helper function to get the optional masp section reference from the Actions. If more than one MaspAction is found we return an error

is_ibc_shielding_transfer

Helper function to check if the action is IBC shielding transfer

Type Aliases

Actions

Actions applied from txs.

Unbond

An unbond of a bond.