namada_sdk::tx
Module data

Data-Types that are used in transactions.

Modules

eval_vp

Tx data to evaluate a validity-predicate

pgf

txs to manage pgf

pos

txs to manage pos Types used for PoS system transactions

protocol

transaction protocols made by validators Types for sending and verifying txs used in Namada protocols

wrapper

wrapper txs

Structs

BatchedTxResult

The result of a specific tx in a batch

DryRunResult

The result of a dry run, included the actual transaction result and the gas used

ExtendedTxResult

The extended transaction result, containing the references to masp sections (if any)

Fee

Amount of some specified token to pay for fees.

GasLimit

Gas limit of a transaction

TxResult

Transaction application result. More specifically the set of inner tx results indexed by the inner tx hash

VpStatusFlags

Validity predicate status flags.

VpsResult

Result of checking a transaction with validity predicates

WrapperTx

A wrapper transaction with some metadata for inclusion and / or verification purposes

Enums

ResultCode

The different result codes that the ledger may send back to a client indicating the status of their submitted tx. The codes must not change with versions, only need ones may be added.

TxSentinel

Sentinel used in transactions to signal events that require special replay protection handling back to the protocol.

TxType

Struct that classifies that kind of Tx based on the contents of its data.

WrapperTxErr

Errors relating to decrypting a wrapper tx and its encrypted payload from a Tx type

Functions

compute_inner_tx_hash

Compute the hash of the some inner tx in a batch.

hash_tx

Get the hash of a transaction

Type Aliases

EllipticCurve