namada_sdk::tx
Module event

Transaction events.

Modules

masp_types

MASP event types

types

Transaction event types.

Structs

Batch

Extend an Event with batch data.

Code

Extend an Event with result code data.

MaspEvent

MASP transaction event

MaspTxRefs

A list of MASP tx references

TxEvent

Transaction event.

Enums

MaspEventKind

MASP event kind

MaspTxRef

A type representing the possible reference to some MASP data, either a masp section or ibc tx data

Functions

new_tx_event

Creates a new event with the hash and height of the transaction already filled in.