namada_sdk
Module uint

An unsigned 256 integer type. Used for, among other things, the backing type of token amounts.

Structs

I256

A signed 256 big integer.

I320

A positive or negative amount

Uint

Little-endian large integer type Namada native type to replace for unsigned 256 bit integers.

Constants

MAX_SIGNED_VALUE

The maximum absolute value a I256 may have. Note the last digit is 2^63 - 1. We add this cap so we can use two’s complement.

MAX_VALUE

The maximum 256 bit integer

ONE

The value one.

ZERO

The value zero.