Crate namada_sdk

Namada SDK

Re-exports

pub use bip39;

pub use masp_primitives;

pub use masp_proofs;

pub use namada_account as account;

pub use namada_gas as gas;

pub use namada_governance as governance;

pub use namada_ibc as ibc;

pub use namada_parameters as parameters;

pub use namada_proof_of_stake as proof_of_stake;

pub use namada_state as state;

pub use namada_storage as storage;

pub use namada_token as token;

pub use tendermint_rpc;

pub use zeroize;

pub use namada_io as io;

pub use namada_wallet as wallet;

Modules

address

Implements transparent addresses as described in Accounts Addresses.

args

Structures encapsulating SDK arguments

arith

Arithmetics helpers

booleans

Boolean related functionality.

borsh

Borsh binary encoding (re-exported) from official crate with custom ext.

bytes

Bytes hex formatting

chain

Chain related data types

collections

Re-exports of collection types.

control_flow

Control flow utilities.

dec

A non-negative fixed precision decimal type.

error

Generic Error Type for all of the Shared Crate

eth_abi

This module defines encoding methods compatible with Ethereum smart contracts.

eth_bridge

Ethereum bridge utilities shared between wasm and the cli.

eth_bridge_pool

The necessary type definitions for the contents of the Ethereum bridge pool

ethereum_events

Types representing data intended for Namada via Ethereum events

ethereum_structs

Ethereum bridge struct re-exports and types to do with ethereum.

events

Logic to do with events emitted by the ledger.

hash

Types for working with 32 bytes hashes.

hints

Compiler hints, to improve the performance of certain operations.

internal

Shared internal types between the host env and guest (wasm).

keccak

This module is for hashing Namada types using the keccak256 hash function in a way that is compatible with smart contracts on Ethereum.

key

Cryptographic keys

masp

MASP verification wrappers.

masp_primitives

General MASP primitives.

queries

Ledger read-only queries can be handled and dispatched via the RPC defined via router! macro.

rpc

SDK RPC queries

signing

Functions to sign transactions

string_encoding

Namada’s standard string encoding for public types.

task_env

Functionality to abstract spawning tasks onto a thread pool.

tendermint

Re-export of tendermint v0.37

tendermint_proto

Re-export of tendermint-proto v0.37

time

Types for dealing with time and durations.

tx

SDK functions to construct different types of transactions

uint

An unsigned 256 integer type. Used for, among other things, the backing type of token amounts.

validity_predicate

Types that are used in validity predicates.

voting_power

This module contains types related with validator voting power calculations.

Macros

impl_display_and_from_str_via_format

Implement std::fmt::Display and std::str::FromStr via Format.

router

Compile time tree patterns router with type-safe dynamic parameter parsing, automatic routing, type-safe path constructors and optional client query methods.

Structs

ExtendedSpendingKey

Wrapper for masp_primitive’s ExtendedSpendingKey

ExtendedViewingKey

Wrapper for masp_primitive’s FullViewingKey

NamadaImpl

Provides convenience methods for common Namada interactions

PaymentAddress

Wrapper for masp_primitive’s PaymentAddress

ShieldedWallet

Represents the current state of the shielded pool from the perspective of the chosen viewing keys.

Enums

DecodeError

TransferSource

Represents a source of funds for a transfer

TransferTarget

Represents a target for the funds of a transfer

Constants

DEFAULT_GAS_LIMIT

Default gas-limit

Traits

MaybeSend

MaybeSync

Namada

An interface for high-level interaction with the Namada SDK

ShieldedUtils

Abstracts platform specific details away from the logic of shielded pool operations.

Functions

decode

Decode a value with borsh

encode

Encode a value with borsh