namada_sdk::eth_bridge
Module bridge_pool

Bridge pool SDK functionality.

Functions

build_bridge_pool_tx

Craft a transaction that adds a transfer to the Ethereum bridge pool.

construct_proof

Construct a merkle proof of a batch of transfers in the bridge pool and return it to the user (as opposed to relaying it to ethereum).

query_bridge_pool

Query the contents of the Ethereum bridge pool. Prints out a json payload.

query_eth_transfer_status

Query the status of a set of transfers to Ethreum, indexed by their keccak hash.

query_relay_progress

Iterates over all ethereum events and returns the amount of voting power backing each TransferToEthereum event.

query_signed_bridge_pool

Query the contents of the Ethereum bridge pool that is covered by the latest signed root. Prints out a json payload.

recommend_batch

Recommend the most economical batch of transfers to relay based on a conversion rate estimates from NAM to ETH and gas usage heuristics.

relay_bridge_pool_proof

Relay a validator set update, signed off for a given epoch.