namada_sdk::key
Trait SigScheme

pub trait SigScheme:
    Eq
    + Ord
    + Debug
    + Serialize
    + Default {
    type Signature: 'static + Signature;
    type PublicKey: 'static + PublicKey;
    type SecretKey: 'static + SecretKey;

    const TYPE: SchemeType;

    // Required methods
    fn generate<R>(csprng: &mut R) -> Self::SecretKey
       where R: CryptoRng + RngCore;
    fn from_bytes(bytes: [u8; 32]) -> Self::SecretKey;
    fn sign_with_hasher<H>(
        keypair: &Self::SecretKey,
        data: impl SignableBytes,
    ) -> Self::Signature
       where H: 'static + StorageHasher;
    fn verify_signature_with_hasher<H>(
        pk: &Self::PublicKey,
        data: &impl SignableBytes,
        sig: &Self::Signature,
    ) -> Result<(), VerifySigError>
       where H: 'static + StorageHasher;

    // Provided methods
    fn sign(
        keypair: &Self::SecretKey,
        data: impl SignableBytes,
    ) -> Self::Signature { ... }
    fn verify_signature(
        pk: &Self::PublicKey,
        data: &impl SignableBytes,
        sig: &Self::Signature,
    ) -> Result<(), VerifySigError> { ... }
}

Represents a digital signature scheme.

More precisely this trait captures the concepts of public keys, private keys, and signatures as well as the algorithms over these concepts to generate keys, sign messages, and verify signatures.

Required Associated Constants

const TYPE: SchemeType

The scheme type of this implementation

Required Associated Types

type Signature: 'static + Signature

Represents the signature for this scheme

type PublicKey: 'static + PublicKey

Represents the public key for this scheme

type SecretKey: 'static + SecretKey

Represents the secret key for this scheme

Required Methods

fn generate<R>(csprng: &mut R) -> Self::SecretKey
where R: CryptoRng + RngCore,

Generate a keypair.

fn from_bytes(bytes: [u8; 32]) -> Self::SecretKey

Instantiate a secret key from the bytes.

fn sign_with_hasher<H>( keypair: &Self::SecretKey, data: impl SignableBytes, ) -> Self::Signature
where H: 'static + StorageHasher,

Sign the data with a key.

fn verify_signature_with_hasher<H>( pk: &Self::PublicKey, data: &impl SignableBytes, sig: &Self::Signature, ) -> Result<(), VerifySigError>
where H: 'static + StorageHasher,

Check that the public key matches the signature on the given data.

Provided Methods

fn sign(keypair: &Self::SecretKey, data: impl SignableBytes) -> Self::Signature

Sign the data with a key, using a SHA256 hasher.

fn verify_signature( pk: &Self::PublicKey, data: &impl SignableBytes, sig: &Self::Signature, ) -> Result<(), VerifySigError>

Check that the public key matches the signature on the given data, using a SHA256 hasher.

Object Safety

This trait is not object safe.