namada_sdk::signing
Function sign_tx

pub async fn sign_tx<'a, D, F, U>(
    wallet: &RwLock<Wallet<U>>,
    args: &Tx,
    tx: &mut Tx,
    signing_data: SigningTxData,
    sign: impl Fn(Tx, PublicKey, Signable, D) -> F,
    user_data: D,
) -> Result<(), Error>
where
    D: Clone + MaybeSend,
    U: WalletIo,
    F: Future<Output = Result<Tx, Error>>,

Sign a transaction with a given signing key or public key of a given signer. If no explicit signer given, use the default. If no default is given, Error.

It also takes a second, optional keypair to sign the wrapper header separately.

If this is not a dry run, the tx is put in a wrapper and returned along with hashes needed for monitoring the tx on chain.

If it is a dry run, it is not put in a wrapper, but returned as is.